### 2024-02-25 #6410 @editorialbot
[PRE REVIEW]: UnfoldSim.jl: A toolbox for simulating continuous event-based time series data for EEG and beyond
**Submitting author:** @jschepers (<a href="http://orcid.org/0009-0000-9270-730X">Judith Schepers</a>)
**Repository:** https://github.com/unfoldtoolbox/UnfoldSim.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v.0.3.1
