### 2024-02-25 #6410 @danielskatz
Author's comment with this submission:

>Dear editors, for transparency reasons, we would like to mention that we submitted an abstract about the UnfoldSim.jl toolbox to the NeuralTraces 2024 workshop for a poster presentation. However, since, to our knowledge, the workshop does not publish any proceedings, this should not count as a publication venue. Moreover, we would like to suggest the following potential reviewers for the paper:
- `@cbrnr` - Clemens Brunner: Julia/EEG toolbox expert
- Adam Wysokiński: NeuroAnalyzer.jl, EEG preprocessing toolbox Julia
- `@lrkrol` - Laurens Krol: SEEREGA, EEG simulation toolbox expert
- `@s4rify`- Sarah Blum: EEG expert, Julia experience
