### 2024-02-25 #6410 @danielskatz
@openjournals/dev - do you have any idea why the reference check isn't happening, either by default when this pre-review was created or now when I try to do it manually?
