### 2024-02-26 #6410 @jschepers
> 👋 @jschepers - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it (it appears to be related to a tab instead of spaces in the metadata). Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Hello 👋,
mh that is weird because it compiles in our repository (with the JOSS Github action). But I looked at the metadata again, replaced any tabs with spaces, and deleted empty lines. Now it works.
I'm sorry for the inconvenience.
