+++ docs/literate/tutorials/multisubject.jl
@@ -1,0 +1,115 @@
+using UnfoldSim
+using Unfold
+using CairoMakie
+using UnfoldMakie
+using DataFrames
+
+# # Multi-Subject simulation
+# Similar to the single subject case, multi-subject simulation depends on:
+# - `Design` (typically a `MultiSubjectDesign`)
+# - `Components` (typically a `MixedModelComponent`)
+# - `Onset` (any)
+# - `Noise` (any)
+
+# ## Design
+
+# Our first design should be 20 subjects, with 3 items each. Any individual image is shown  only either as large or small, thus we choose `items_between`. 
+design = MultiSubjectDesign(
+    n_subjects = 20,
+    n_items = 4,
+    items_between = Dict(:condition => ["large", "small"]),
+)
+
+
+# ## Between, within?
+# In the beginning, the distinction between `between-items`, `between-subjects` and `within-subjects`, `within-items` and `both-between`, `both-within` feels daunting.
+# 
+# We base our terminology on `MixedModelsSim` which uses the following definitions:
+# - `subjects_between` -> effects between subjects, e.g. young vs old
+# - `items_between` -> effects between items, e.g. natural vs artificial images, (but shown to all subjects if not specified in subjects_between as well)
+# - `both_within` -> effects completly crossed, e.g. word vs. scramble, where the "original" word is the item, and shown to all subjects 
+
+
+# ## Components
+# For multi-subject, similar to the `LinearModelComponent` specified before, we have to define the fixed effect `β`, the model parameters that are applied to all subjects.
+β = [1, 2] # 1 = intercept, 2 = condition_is_large effect
+
+# In addition, we have to provide random effects `σs`, which define the spread (and  correlation) of the subjects around the fixed effects, foreach parameter
+σs = Dict(
+    :subject => [0.5, 1], # we have more spread in the condition-effect
+    :item => [1], # the item-variability is higher than the subject-variability 
+)
+
+# now we are ready to build it together
+signal = MixedModelComponent(;
+    basis = UnfoldSim.hanning(50),
+    formula = @formula(0 ~ 1 + condition + (1 + condition | subject) + (1 | item)),
+    β = β,
+    σs = σs,
+    contrasts = Dict(:condition => EffectsCoding()), # we highly recommend specifying your contrasts, by Default its Dummy/ReferenceCoding with alphabetically sorted levels (relying 100% on StatsModels.jl)
+)
+
+# and simulate!
+data, evts = simulate(design, signal, NoOnset(), NoNoise(), return_epoched = true);
+
+# We get data with 50 samples (our `basis` from above), with `2` trials/items and 20 subjects. We get items and subjects separately because we chose no-overlap (via `NoOnset`) and `return_epoched=true``.
+@info size(data)
+first(evts, 5)
+
+# Finally, let's plot the data
+f = Figure()
+
+for k = 1:4
+    series(
+        f[1, k],
+        data[:, k, :]',
+        solid_color = :black,
+        axis = (; limits = ((0, 50), (-4, 5))),
+    )
+    Label(f[1, k, Top()], text = "Item:" * evts[k, :item] * ", c:" * evts[k, :condition])
+end
+f
+
+# Some remarks on interpreting the plot:
+# - The β main-effect of small / Large (#1 and #3 vs. #2 and #4) is clearly visible. 
+# - The variability between subjects, is the variability between the individual curves.
+# - The item effect shows up e.g. that #2 vs. #4 column show different values.
+
+
+# # Continuous Signals / Overlap
+# Let's continue our tutorial and simulate overlapping signals instead.
+#
[…]
