+++ scripts/sketch.jl
@@ -1,14 +1,0 @@
-using MixedModelsSim
-using MixedModels
-using Random
-
-
-
-
-dfA = get_df(1)
-dfB = get_df(2)
-
-
-df = dfA # for debug
-
-signal_coef(dfA)
