### 2024-02-26 c46643c
fix italic-issue in docstring
