### 2024-02-26 ea7a7c7
wadded warning for extreme inter-onset-distances
