### 2024-02-26 5d4f0a0
better docstring + enforce Dict{Symbol}
