### 2024-02-26 e085877
remove creation for Simulation object in tutorial
