### 2024-02-26 7690388
Set minimum for first onset to 1 (not 0)
