### 2024-02-26 5128e3f
Set minimum for first onset to 1 (not 0)
