### 2024-02-26 #6410 @danielskatz
👋 @mooniean - do you think you would be able to edit this JOSS submission?  Note that the author has already suggested some reviewers [above](https://github.com/openjournals/joss-reviews/issues/6410#issuecomment-1963015879).
