+++ joss_paper/paper.bib
@@ -8,7 +8,7 @@
   pages={e7838},
   year={2019},
   publisher={PeerJ Inc.},
-  doi = {https://doi.org/10.7717/peerj.7838}
+  doi = {10.7717/peerj.7838}
 }
 
 @article{dimigen2021regression,
@@ -316,7 +316,7 @@ url = {https://doi.org/10.5281/zenodo.8344531}
   year={2021},
   publisher={Elsevier},
   issn = {0165-0270},
-  doi = {https://doi.org/10.1016/j.jneumeth.2020.109017},
+  doi = {10.1016/j.jneumeth.2020.109017},
   url = {https://www.sciencedirect.com/science/article/pii/S0165027020304404},
 }
 
@@ -344,7 +344,7 @@ url = {https://doi.org/10.5281/zenodo.8344531}
   year={2019},
   publisher={Elsevier},
   issn = {0165-0270},
-  doi = {https://doi.org/10.1016/j.jneumeth.2019.108377},
+  doi = {10.1016/j.jneumeth.2019.108377},
   url = {https://www.sciencedirect.com/science/article/pii/S0165027019302341},
 }
 
@@ -371,7 +371,7 @@ url = {https://doi.org/10.5281/zenodo.8344531}
   year={2018},
   publisher={Elsevier},
   issn = {0165-0270},
-  doi = {https://doi.org/10.1016/j.jneumeth.2020.109017},
+  doi = {10.1016/j.jneumeth.2020.109017},
   url = {https://www.sciencedirect.com/science/article/pii/S0165027020304404},
 }
 
