### 2024-02-26 93cde2a
Apply suggestions from code review
Co-authored-by: Judith Schepers <judith.schepers@vis.uni-stuttgart.de>
