### 2024-02-26 e9ff6d6
Apply suggestions from code review
Co-authored-by: Judith Schepers <judith.schepers@vis.uni-stuttgart.de>
