### 2024-02-27 1bab9dd
Update docs/literate/tutorials/multisubject.jl
Co-authored-by: Judith Schepers <judith.schepers@vis.uni-stuttgart.de>
