### 2024-02-27 #6410 @jschepers
> I'll also point out that your paper is about twice as long as JOSS suggests (~1900 words vs 250-1000). You might want to consider if there's anything you can remove in favor of pointing at other documents in the repo. If reviewers ask for more to be added, you will certainly have to remove something at the same time so this doesn't get any longer.

Thank you for pointing this out. We will start thinking about this and keep it in mind during the review process.
