### 2024-02-28 6c35f3c
half way through to success for sequence designs or something
