+++ src/component.jl
@@ -27,6 +27,7 @@ MixedModelComponent(;
     β::Vector
     σs::Dict # Dict(:subject=>[0.5,0.4]) or to specify correlationmatrix Dict(:subject=>[0.5,0.4,I(2,2)],...)
     contrasts::Dict = Dict()
+    offset::Int = 0
 end
 
 """
@@ -55,7 +56,24 @@ LinearModelComponent(;
     formula::Any # e.g. 0~1+cond - left side must be "0"
     β::Vector
     contrasts::Dict = Dict()
+    offset::Int = 0
 end
+
+
+"""
+    offset(AbstractComponent)
+
+Should the `basis` be shifted? Returns c.offset for most components, if not implemented for a type, returns 0. Can be positive or negative, but has to be Integer
+"""
+offset(c::AbstractComponent)::Int = 0
+offset(c::LinearModelComponent)::Int = c.offset
+offset(c::MixedModelComponent)::Int = c.offset
+
+maxoffset(c::Vector{<:AbstractComponent}) = maximum(offset.(c))
+maxoffset(d::Dict{<:Char,<:Vector{<:AbstractComponent}}) = maximum(maxoffset.(values(d)))
+minoffset(c::Vector{<:AbstractComponent}) = minimum(offset.(c))
+minoffset(d::Dict{<:Char,<:Vector{<:AbstractComponent}}) = minimum(minoffset.(values(d)))
+
 
 
 """
@@ -93,6 +111,7 @@ For `MultichannelComponent` return the length of the projection vector.
 
 """
 n_channels(c::MultichannelComponent) = length(c.projection)
+
 
 """
 For a vector of `MultichannelComponent`s, return the first but asserts all are of equal length.
@@ -145,7 +164,7 @@ simulate_component(rng, c::AbstractComponent, simulation::Simulation) =
     simulate_component(rng, c, simulation.design)
 
 """
-    simulate_component(rng, c::AbstractComponent, simulation::Simulation)
+    simulate_component(rng, c::LinearModelComponent, design::AbstractDesign)
 Generate a linear model design matrix, weight it by c.β and multiply the result with the given basis vector.
 
 julia> c = UnfoldSim.LinearModelComponent([0,1,1,0],@formula(0~1+cond),[1,2],Dict())
@@ -192,13 +211,13 @@ Currently, it is not possible to use a different basis for fixed and random effe
 
 julia> design = MultiSubjectDesign(;n_subjects=2,n_items=50,items_between=(;:cond=>["A","B"]))
 julia> c = UnfoldSim.MixedModelComponent([0.,1,1,0],@formula(0~1+cond+(1|subject)),[1,2],Dict(:subject=>[2],),Dict())
-julia> simulate(StableRNG(1),c,design)
+julia> simulate_component(StableRNG(1),c,design)
 
 """
 function simulate_component(
     rng,
     c::MixedModelComponent,
-    design::AbstractDesign;
+    design::AbstractDesign,
     return_parameters = false,
 )
     events = generate_events(design)
@@ -335,34 +354,51 @@ function simulate_responses!(
     return epoch_data
 end
 function init_epoch_data(components, design)
+    max_offset = maxoffset(components)
+    min_offset = minoffset(components)
+    range_offset = (max_offset - min_offset)
     if n_channels(components) > 1
-        epoch_data = zeros(n_channels(components), maxlength(components), length(design))
+        epoch_data = zeros(
+            n_channels(components),
+            maxlength(components) + range_offset,
+            length(design),
+        )
     else
-        epoch_data = zeros(maxlength(components), length(design))
+        epoch_data = zeros(maxlength(components) + range_offset, length(design))
     end
     return epoch_data
 end
 
[…]
