### 2024-02-28 809dc2e
everythinig except sequencelength seems to be working now
