+++ src/simulation.jl
@@ -29,6 +29,8 @@ Some remarks to how the noise is added:
   - If `return_epoched = true` and `onset =NoOnset()` the noise is added to the epoched data matrix
   - If `onset` is not `NoOnset`, a continuous signal is created and the noise is added to this i.e. this means that the noise won't be the same as in the `onset = NoOnset()` case even if `return_epoched = true`.
   - The case `return_epoched = false` and `onset = NoOnset()` is not possible and therefore covered by an assert statement
+  - `simulate(rng,design::SequenceDesign,...)`
+    If no `design.rng` was defined for `SequenceDesign`, we replace it with the `simulation`-function call `rng` object
 
 """
 
@@ -38,29 +40,33 @@ function simulate(args...; kwargs...)
     simulate(MersenneTwister(1), args...; kwargs...)
 end
 
-simulate(
+function simulate(
     rng::AbstractRNG,
     design::AbstractDesign,
     signal,
     onset::AbstractOnset,
     noise::AbstractNoise = NoNoise();
     kwargs...,
-) = simulate(rng, Simulation(design, signal, onset, noise); kwargs...)
-
-function simulate(
-    rng::AbstractRNG,
-    design::SequenceDesign,
-    signal,
-    onset::AbstractOnset,
-    noise::AbstractNoise = NoNoise();
-    kwargs...,
 )
-    design =
-        isnothing(design.rng) ?
-        SequenceDesign(design.design, design.sequence, design.sequencelength, rng) : design
+
+    if is_SequenceDesign(design)
+        design = sequencedesign_add_rng(rng, design)
+    end
     simulate(rng, Simulation(design, signal, onset, noise); kwargs...)
 end
 
+
+sequencedesign_add_rng(rng, design::AbstractDesign) = design
+sequencedesign_add_rng(rng, design::RepeatDesign) =
+    RepeatDesign(sequencedesign_add_rng(rng, design.design), design.repeat)
+sequencedesign_add_rng(rng, design::SequenceDesign) =
+    isnothing(design.rng) ?
+    SequenceDesign(design.design, design.sequence, design.sequencelength, rng) : design
+
+
+is_SequenceDesign(d::AbstractDesign) = false
+is_SequenceDesign(d::RepeatDesign) = is_SequenceDesign(d.design)
+is_SequenceDesign(d::SequenceDesign) = true
 
 
 function simulate(rng::AbstractRNG, simulation::Simulation; return_epoched::Bool = false)
@@ -191,8 +197,8 @@ function add_responses!(signal, responses::Vector, e, s, tvec, erpvec)
     @views signal[e, tvec, s] .+= responses[:, erpvec]
 end
 function add_responses!(signal, responses::Matrix, e, s, tvec, erpvec)#
-    @debug size(signal), size(responses), e, s, size(tvec), size(erpvec)
-    @debug tvec, erpvec
+    #    @debug size(signal), size(responses), e, s, size(tvec), size(erpvec)
+    #@debug tvec, erpvec
     @views signal[e, tvec, s] .+= responses[:, erpvec]
 end
 function add_responses!(signal, responses::AbstractArray, e, s, tvec, erpvec)
