### 2024-03-07 #6410 @mooniean
Hi 👋 @cbrnr  @lrkrol  @s4rify - would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> UnfoldSim.jl: A toolbox for simulating continuous event-based time series data for EEG and beyond #6410

Thanks in advance!
