### 2024-03-10 b4a8ac8
generalized LinearModelComponent to arbitrary functions instead of vectors
