+++ src/simulation.jl
@@ -35,9 +35,15 @@ Some remarks to how the noise is added:
 """
 
 
-function simulate(args...; kwargs...)
+function simulate(
+    design::AbstractDesign,
+    signal,
+    onset::AbstractOnset,
+    noise::AbstractNoise;
+    kwargs...,
+)
     @warn "No random generator defined, used the default (`Random.MersenneTwister(1)`) with a fixed seed. This will always return the same results and the user is strongly encouraged to provide their own random generator!"
-    simulate(MersenneTwister(1), args...; kwargs...)
+    simulate(MersenneTwister(1), design, signal, onset, noise; kwargs...)
 end
 
 function simulate(
