### 2024-03-11 57a7f68
forgot to define offset in LinearModelFunction
