+++ src/predefinedSimulations.jl
@@ -46,9 +46,9 @@ function predef_eeg(
 
     # component / signal
     sfreq = 100,
-    p1 = (p100(; sfreq = sfreq), @formula(0 ~ 1), [5], Dict()),
-    n1 = (n170(; sfreq = sfreq), @formula(0 ~ 1 + condition), [5, -3], Dict()),
-    p3 = (p300(; sfreq = sfreq), @formula(0 ~ 1 + continuous), [5, 1], Dict()),
+    p1 = (p100(; sfreq = sfreq), @formula(0 ~ 1), [5], Dict(), 0),
+    n1 = (n170(; sfreq = sfreq), @formula(0 ~ 1 + condition), [5, -3], Dict(), 0),
+    p3 = (p300(; sfreq = sfreq), @formula(0 ~ 1 + continuous), [5, 1], Dict(), 0),
     kwargs...,
 )
 
