### 2024-03-15 #6410 @mooniean
That's great! Let's give it a shot :)

Hey @agricolab @ahwillia @vankesteren! Would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> UnfoldSim.jl: A toolbox for simulating continuous event-based time series data for EEG and beyond #6410

Thanks in advance!
