### 2024-03-15 #6410 @ahwillia
I'm unavailable for the next few months.

On Fri, Mar 15, 2024, 9:03 AM Beatriz Costa Gomes ***@***.***>
wrote:

> That's great! Let's give it a shot :)
>
> Hey @agricolab <https://github.com/agricolab> @ahwillia
> <https://github.com/ahwillia> @vankesteren
> <https://github.com/vankesteren>! Would you be available and interested
> in reviewing this submission for the Journal of Open Source Software?
>
> UnfoldSim.jl: A toolbox for simulating continuous event-based time series
> data for EEG and beyond #6410
> <https://github.com/openjournals/joss-reviews/issues/6410>
>
> Thanks in advance!
>
