### 2024-03-21 #6410 @mooniean
Hi @vahid-sb @behinger @pitmonticone - would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> UnfoldSim.jl: A toolbox for simulating continuous event-based time series data for EEG and beyond #6410

Thanks in advance!
