### 2024-03-21 #6410 @jschepers
Thank you everyone for your answers and thank you @mooniean for trying to find reviewers. @behinger is the senior author of the paper in question. But we thought one could ask AdamWysokinski or marcpabst  (although we don't know whether they are on the list of JOSS reviewers).
