### 2024-03-22 #6410 @behinger
(for transparency reasons: marcpabst and I are started few weeks back to exchange ideas on an AbstractEEG-interface open source project - I don't know if JOSS sees this as a conflict of interest)
