### 2024-04-11 #6410 @mooniean
> (for transparency reasons: marcpabst and I are started few weeks back to exchange ideas on an AbstractEEG-interface open source project - I don't know if JOSS sees this as a conflict of interest)

Hello! Thank you for flagging this! If it is a collaboration, then it is a COI.

I'll try and find some Julia reviewers for this paper!
