### 2024-04-11 #6410 @mooniean
Hi @roualdes @aurorarossi @jonbmartin - would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> UnfoldSim.jl: A toolbox for simulating continuous event-based time series data for EEG and beyond #6410

Thanks in advance!
