### 2024-04-11 #6410 @jonbmartin
Hi Beatriz,

Looks very interesting, but I am not able to at the moment.

Jonathan

On Thu, Apr 11, 2024 at 3:32 AM Beatriz Costa Gomes <
***@***.***> wrote:

> Hi @roualdes <https://github.com/roualdes> @aurorarossi
> <https://github.com/aurorarossi> @jonbmartin
> <https://github.com/jonbmartin> - would you be available and interested
> in reviewing this submission for the Journal of Open Source Software?
>
> UnfoldSim.jl: A toolbox for simulating continuous event-based time series
> data for EEG and beyond #6410
> <https://github.com/openjournals/joss-reviews/issues/6410>
>
> Thanks in advance!
>
