+++ docs/nb_muscle.jl
@@ -1,0 +1,2326 @@
+### A Pluto.jl notebook ###
+# v0.19.37
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 080f23f4-f900-11ee-1e5d-cd486a86ccff
+begin
+	using UnfoldSim
+	using DSP
+	using CairoMakie
+	using Random
+	using Parameters
+	using UnfoldMakie
+	using DataFrames
+end
+	
+
+# ╔═╡ 5d925343-5d02-4887-9337-11cdf810c698
+hart = headmodel()
+
+# ╔═╡ bf4d7bdf-77df-402e-8c4e-098e85fcc34a
+T = 10; # total time
+
+# ╔═╡ b4beaaa3-7f98-4f12-a8d9-38994974e462
+srate = 100; #sampling rate
+
+# ╔═╡ b53f298d-e420-44b1-b1ce-cd5f8fe514d0
+musclelabels = unique(hart.artefactual["label"])
+
+# ╔═╡ f56aa7db-4dbd-404f-8704-3328b190b0ca
+what_muscle = rand(musclelabels,1)
+
+# ╔═╡ 6352a797-eb24-4964-9c34-0d410310df4e
+# get a "random" muscle & a random muscle-sourcepoint
+lead_ix = rand(findall(hart.artefactual["label"].==what_muscle),1)[1]
+
+# ╔═╡ 17c5239c-04a0-47cd-9393-115763e7aaad
+lead = magnitude(hart.artefactual["leadfield"][:,lead_ix:lead_ix,:],hart.artefactual["orientation"][lead_ix:lead_ix,:]);
+
+# ╔═╡ 76575e4c-df59-49a3-8799-0aa0db4d35db
+design = SingleSubjectDesign(conditions=Dict(:a=>["b"]))
+
+# ╔═╡ d01bbf66-1f7e-4646-b8e4-4d6ec6539533
+@with_kw struct MuscleNoise <: AbstractNoise
+	srate
+    noiselevel = 1
+end
+
+# ╔═╡ a005fef0-b145-492b-88d7-198ba1a751c9
+function UnfoldSim.simulate_noise(rng, t::MuscleNoise, n::Int)
+	x = rand(rng,n).-0.5
+	
+	fi = digitalfilter(Bandstop(5,20; fs=t.srate),Butterworth(2)) # this kind of gives us the sqrt-shape in FFT space
+	return t.noiselevel .* filt(fi,x)
+end
+
+
+# ╔═╡ 796eb40e-2e70-492e-bca1-af197519684d
+begin
+c = LinearModelComponent(; basis = zeros(T*srate), formula = @formula(0 ~ 1), β = [1]); # dummy LinearModelComponent
+mc = UnfoldSim.MultichannelComponent(c, lead[:,1],MuscleNoise(srate=srate)) # Multichannelcomponent with MuscleNoise + Leadfield
+
+end
+
+# ╔═╡ 9248e9d0-2fc6-4e82-8608-3c259b8a3c86
+data, events = simulate(MersenneTwister(1),design, mc, NoOnset(), NoNoise();return_epoched=true);
+
+
+# ╔═╡ c7991ae2-27f3-4d96-a667-d1e6130d01a3
+begin
+	pos3d = hart.electrodes["pos"];
+	pos2d = to_positions(pos3d')
+	pos2d = [Point2f(p[1] + 0.5, p[2] + 0.5) for p in pos2d];
+end
+
+# ╔═╡ 9821d37d-007d-497e-83e7-4f324696e8a2
+let
+	# select only every 20th channel
+	n_th = 30
[…]
