### 2024-04-17 #6641 @editorialbot
[REVIEW]: UnfoldSim.jl: A toolbox for simulating continuous event-based time series data for EEG and beyond
**Submitting author:** @jschepers (<a href="http://orcid.org/0009-0000-9270-730X">Judith Schepers</a>)
**Repository:** https://github.com/unfoldtoolbox/UnfoldSim.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.4.0

@jonbmartin & @cbrnr & @aurorarossi & @roualdes, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
