### 2024-04-18 #6641 @aurorarossi
@jschepers

- [ ]  The word count of the paper is **1984**. As written in the [JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) where it says *"The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper"*. I would suggest to remove the Package References paragraph and to significantly reduce the Simulation Example paragraph, since this can also be found in the package documentation.

- [ ]  You mention contributing in the `README.md`, but I suggest adding a `CONTRIBUTING.md` file as it is more structured and good practice, as in the example below:

    https://github.com/JuliaGraphs/Graphs.jl/blob/master/CONTRIBUTING.md
    https://github.com/FluxML/Flux.jl/blob/master/CONTRIBUTING.md
    https://github.com/JuliaGraphs/GraphsOptim.jl/blob/main/CONTRIBUTING.md
