### 2024-04-25 #6641 @cbrnr
I'm struggling a bit with the following point:

The first part is pretty clear, the submitting authors has definitely made major contributions to the software. However, I'm not sure that this is also the case for the second author @llips, who has only one trivial commit (see https://github.com/unfoldtoolbox/UnfoldSim.jl/graphs/contributors). Since there is no author contributions section, I would argue that this author should probably not be on the authors list based on the information I have. But I'm happy to hear any clarification of course!
