### 2024-04-26 #6641 @cbrnr
Regarding the paper, I agree with @aurorarossi that it is currently too long. I wouldn't reduce the Simulation chapter though, because although it can be found in the documentation, this is the gist of the entire package. I think a few sentences/paragraphs can be removed without too much negative effects throughout the entire manuscript. I would especially try to shorten the validation part, which, although very interesting, is not the focus of this package, and at least for me it raises several questions that are better discussed elsewhere. I agree that the package references can be completely removed (and if you want to credit these packages, maybe just one sentence pointing to some part of the docs would be sufficient). Also, the "Related tools" section could be considerably shortened.

I also have a couple of suggestions that I think would improve the paper:

- [ ] _Title_: "Toolbox" has strong MATLAB vibes, I suggest to avoid this term and use e.g. "package" instead. My suggestion for an improved title is: "UnfoldSim.jl: Simulating continuous event-based time series data for EEG and beyond"
- [ ] _Statement of need_: "For a multitude of reasons, we **often** need to simulate such kind of data: Simulated EEG data is ~~necessary~~**useful** to test preprocessing and analysis tools, ..."
- [ ] At the end of the "Statement of need" section, you could mention how UnfoldSim.jl overcomes each of these limitations (in a short sentence)? Otherwise, it is not really explicit why being MATLAB-based is a limitation (although most readers will probably know it of course).
- [ ] _Figure 1 (upper panel)_: Parameters in the legend are on a log scale, but the plot shows distributions on a linear scale. This makes it a bit difficult to relate the selected parameters to the distributions. I suggest to convert the parameters in the legend accordingly.
- [ ] _Figure 1 (lower panel)_: For the uniform distribution, it would be more intuitive to mention the offset before the width in the legend.
- [ ] _Figure 2 A_: Why is there a negative offset for `ExponentialNoise`? I assume the reason is to avoid overlaps, which is a good idea, but then I would add a custom offset to each noise type to show each time series without any overlap.
- [ ] _Figure 2 B_: Why is `NoNoise` in the legend (clearly, it is not shown in the PSD)?
- [ ] _Figure 2 caption_ (nitpick): log and power should not be italicized.
- [ ] _Simulation example, step 1_: "continuous variable with 10 levels" sounds like you describe a factor. I guess the term "levels" should be replaced here with e.g. "values" or something like that.
- [ ] _All code snippets_: Is it really necessary to suppress output with `;`? If not, I'd rather not see this throughout the entire manuscript (you know, MATLAB vibes again)...
- [ ] _Table 1_: Where does "latency" come from and why does it have the specific values shown here?
- [ ] _Simulation example, step 2_: Just wondering, why has the basis to be specified as a function call as opposed to just passing the function?
- [ ] _Simulation example, step 2_: Maybe relate these specific parameters 5 and 3 to the desired component(s)?
- [ ] I recommend to shorten the validation part, but if you are going to keep the code snippet, then briefly explain `UnfoldModel`.
- [ ] Final nitpick: please use a ' instead of ´ in "Germany's Excellence Strategy".

Once these issues are resolved, I think the paper is good to go and I will focus on the software.
