+++ previews/PR94/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"EditURL = \"../../../literate/HowTo/newComponent.jl\"","category":"page"},{"location":"generated/HowTo/newComponent/#New-component:-Duration-Shift","page":"Define a new duration & jitter component","title":"New component: Duration + Shift","text":"","category":"section"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"We want a new component that changes its duration and shift depending on a column in the event-design. This is somewhat already implemented in the HRF + Pupil bases","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"using UnfoldSim\nusing Unfold\nusing Random\nusing DSP\nusing CairoMakie, UnfoldMakie\n\nsfreq = 100;\nnothing #hide","category":"page"},{"location":"generated/HowTo/newComponent/#Design","page":"Define a new duration & jitter component","title":"Design","text":"","category":"section"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"Let's generate a design with two columns, shift + duration","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"design = UnfoldSim.SingleSubjectDesign(;\n    conditions = Dict(\n        :shift => rand(100) .* sfreq / 5,\n        :duration => 20 .+ rand(100) .* sfreq / 5,\n    ),\n)","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"We also need a new AbstractComponent","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"struct TimeVaryingComponent <: AbstractComponent\n    basisfunction::Any\n    maxlength::Any\nend","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"We have to define the length of a component","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"Base.length(c::TimeVaryingComponent) = length(c.maxlength)","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"While we could have put the TimeVaryingComponent.basisfunction directly into the simulate function, I thought this is a bit more modular","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"function UnfoldSim.simulate(rng, c::TimeVaryingComponent, design::AbstractDesign)\n    evts = generate_events(design)\n    return c.basisfunction(evts, c.maxlength)\nend","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new duration & jitter component","title":"Define a new duration & jitter component","text":"finally, the actual function that does the shifting + duration","categ
[…]
