### 2024-05-14 #6641 @jonbmartin
Hello Beatriz,

I think there may have been a miscommunication - I initially said that I
was unable to review. I am currently working on some other reviews for JOSS
and other journals. I will be happy to review in the future!

Very best,
Jonathan Martin

On Wed, May 8, 2024 at 5:32 AM Beatriz Costa Gomes ***@***.***>
wrote:

> Hi @jschepers <https://github.com/jschepers> and reviewers @cbrnr
> <https://github.com/cbrnr> @roualdes <https://github.com/roualdes>
> @aurorarossi <https://github.com/aurorarossi> and @jonbmartin
> <https://github.com/jonbmartin> ! Just checking in how things are going!
>
