### 2024-05-14 unfoldtoolbox/UnfoldSim.jl#95 @roualdes
JOSS review: doc suggestions
The doc for UnfoldSim.jl is doing great.  Lots of good stuff in there to help users navigate the functionality and explore the API.

Here's some comments/suggestions that I think could further improve the experience for new users.

* Quickstart page
    *  This page leads with a tip that is devoid of context.  It's hard for me to understand how this tip is intended to be used.
    *  Only in sub-section Combine & Generate does one learn what we've been working towards.  Maybe there could be an introductory paragraph or two that explains the big picture idea of what's going on with the multiple components and explains how they will be combined to generated simulated data.
* Simulate ERPs page
    * An introductory paragraph seems like it could be useful to explain the context going on in the authors' mind.  For instance, the quoted sentence fragment "faces are more negative than cars" suggests that an example with very specific context exists.  However, this context is not laid out for the user. Maybe I'm just slow on the uptake here, because this isn't directly my field of study.
* Poweranalysis page
    * There is a good opening sentence to explain what's going on.  Yay.
    * Maybe Power Analysis, two words, instead of one.
* Reference Section
    * From the README, "All components (design, components, onsets, noise) can be easily modified and you can simply plugin your own!"  The word component is overloaded in a way that doesn't help me keep the ideas of the package in my head.  The Reference section is seemingly describing the details of the components necessary for a simulation.  But one of the components is called component.
    * The sub-titles found in the sidebar of this section do not match the titles of the pages; e.g. Overview: Toolbox Functions corresponds to Overview of functionality, and Overview components doesn't have a title on the page, only sub-headers.
    * The order of the sub-titles as described on the first Overview page doesn't match the order of the sub-sections.  The first overview page provides sub-headers: Design, Components, Onset, Noise.  But the subsections go as: Noise, Onset, Components.  Further, there is no Design page.
* API/DocStrings
    * `simulate()` appears to be missing an entry
    * `SingleSubjectDesign()` could use a sentence or two explaining the basic idea of the function
    * Something odd seems to be going on for `predef_eeg()`
    * Do people in this field know without reference the different colors of noise?  I think a quick link could be helpful, but I'm not in this field
    * `RealisticNoise()` doc misspells Artifact.jl
    * `ExponentialNoise()` contains a warning, but it seems like a helpful message letting the user know what `N` refers to in your package could help users know better what to avoid.  Is it memory or computational cost that you're encouraging users to avoid? The Cholesky decomposition scales computationally as $O(N^3)$, where $N$ is the number of rows (or columns, since square), but I less often see statements about the memory costs.

Over all, I'd suggesting making titles (in the sidebar) match better to the main header of each page and adding at least an introductory sentence or paragraph to each page and function, to let the user know what they're about to read about.
