### 2024-05-15 #6641 @jschepers
> I filed a new issue, [unfoldtoolbox/UnfoldSim.jl#95](https://github.com/unfoldtoolbox/UnfoldSim.jl/issues/95), with all my doc comments and suggestions. I'll wait for updates on the paper before I read that over.

Thanks a lot for your detailed comments on the documentation. That's very helpful.
