### 2024-07-10 d2d3e7b
Improve documentation especially quickstart, home and power analysis
