+++ docs/src/index.md
@@ -2,14 +2,14 @@
 CurrentModule = UnfoldSim
 ```
 
-# UnfoldSim
+# UnfoldSim.jl
 
-Documentation for [UnfoldSim](https://github.com/behinger/UnfoldSim.jl).
+Documentation for [UnfoldSim.jl](https://github.com/unfoldtoolbox/UnfoldSim.jl). UnfoldSim.jl is a Julia package for simulating multivariate timeseries data with a special focus on EEG data.
 
 ## Start simulating timeseries
-We offer some predefined signals, check them out!
+We offer some predefined (EEG) signals, check them out!
 
-For instance an P1/N170/P300 complex.
+For instance a P1/N170/P300 complex (containing three typical ERP components).
 ```@example main
 using UnfoldSim
 using CairoMakie # plotting
