### 2024-07-10 85f037b
Improve documentation especially quickstart, home and power analysis
