### 2024-07-23 b073526
added draft of design types reference page
