### 2024-07-24 9ab04aa
Finished experimental design reference page
