### 2024-07-30 a018481
Improved docstrings for single- and multi-subject design
