### 2024-07-31 bb74b94
Added cross references in docstrings
