### 2024-07-31 f9424a5
Added intro sentences, matched titles and sidebar, reordered pages and added collapsible setup blocks
