### 2024-07-31 541cf67
Added intro sentences, matched titles and sidebar, reordered pages and added collapsible setup blocks
