+++ docs/literate/reference/overview.jl
@@ -1,7 +1,17 @@
 # # Overview of functionality
 # UnfoldSim has many modules, here we try to collect them to provide you with an overview.
+
+# ### Setup
+# ```@raw html
+# <details>
+# <summary>Click to expand</summary>
+# ```
+## Load required packages
 using UnfoldSim
 using InteractiveUtils
+# ```@raw html
+# </details>
+# ```
 
 # ## Design
 # Designs define the experimental design. They can be nested, e.g. `RepeatDesign(SingleSubjectDesign,10)` would repeat the generated design-dataframe 10x.
