+++ docs/make.jl
@@ -19,7 +19,7 @@ DocMeta.setdocmeta!(UnfoldSim, :DocTestSetup, :(using UnfoldSim); recursive = tr
 
 makedocs(;
     modules = [UnfoldSim],
-    authors = "Luis Lips, Benedikt Ehinger, Judith Schepers",
+    authors = "Judith Schepers, Luis Lips, Maanik Marathe, Benedikt Ehinger",
     #repo="https://github.com/unfoldtoolbox/UnfoldSim.jl/blob/{commit}{path}#{line}",
     repo = Documenter.Remotes.GitHub("unfoldtoolbox", "UnfoldSim.jl"),
     sitename = "UnfoldSim.jl",
@@ -27,13 +27,14 @@ makedocs(;
         prettyurls = get(ENV, "CI", "false") == "true",
         canonical = "https://unfoldtoolbox.github.io/UnfoldSim.jl",
         edit_link = "main",
+        sidebar_sitename = false,
         assets = String[],
     ),
     pages = [
         "Home" => "index.md",
         "Tutorials" => [
             "Quickstart" => "generated/tutorials/quickstart.md",
-            "Simulate ERPs" => "generated/tutorials/simulateERP.md",
+            "Simulate event-related potentials (ERPs)" => "generated/tutorials/simulateERP.md",
             "Power analysis" => "generated/tutorials/poweranalysis.md",
             "Multi-subject simulation" => "generated/tutorials/multisubject.md",
         ],
@@ -45,13 +46,13 @@ makedocs(;
             "Overview: Noise types" => "./generated/reference/noisetypes.md",
         ],
         "HowTo" => [
-            "Define a new, (imbalanced) design" => "./generated/HowTo/newDesign.md",
-            "Repeating a design" => "./generated/HowTo/repeatTrials.md",
-            "Define a new duration & jitter component" => "./generated/HowTo/newComponent.md",
+            "Define a new (imbalanced) design" => "./generated/HowTo/newDesign.md",
+            "Get multiple trials with identical subject/item combinations" => "./generated/HowTo/repeatTrials.md",
+            "Define a new component (with variable duration and shift)" => "./generated/HowTo/newComponent.md",
             "Generate multi channel data" => "./generated/HowTo/multichannel.md",
-            "Use predefined design / onsets data" => "./generated/HowTo/predefinedData.md",
+            "Use existing experimental designs & onsets in the simulation" => "./generated/HowTo/predefinedData.md",
         ],
-        "API / DocStrings" => "api.md",
+        "API / Docstrings" => "api.md",
     ],
 )
 
