### 2024-07-31 aca3fd2
Update src/noise.jl
