### 2024-07-31 fb05406
removed NoNoise from the noise comparison plot
