### 2024-07-31 b6a9d88
Start addressing JOSS review comments
