### 2024-08-01 43dd57c
Update docs/literate/reference/noisetypes.jl
Co-authored-by: Benedikt Ehinger <benedikt.ehinger@vis.uni-stuttgart.de>
