### 2024-08-01 2c34e5e
Update docs/literate/reference/noisetypes.jl
