### 2024-08-05 a6c049e
Adapt noise figure and caption
