### 2024-08-05 ae032b5
adapted figure size
