### 2024-08-05 c966897
adapted figure size
