### 2024-08-05 783be2f
adapted figure size
