### 2024-08-05 bc57c9b
adapted figure size
