### 2024-08-05 9bf8d6b
adapted figure size
