### 2024-08-05 ff86cda
adapted figure size
