### 2024-08-05 1f80832
adapted figure size
