### 2024-08-05 e96ef9c
adapted figure size
