### 2024-08-05 c46f58f
adapted figure size
