### 2024-08-05 74985f1
adapted figure size
