### 2024-08-06 #6641 @jschepers
Dear all,
First of all, we thank you for being patient with our late reply. Thanks a lot again for all your helpful feedback on the paper and the package. Please find the changes we made and the additional clarifications in separate posts below.
Benedikt & Judith
