### 2024-08-06 #6641 @jschepers
Dear @aurorarossi,
thanks a lot for your suggestion about adding a `CONTRIBUTING.md` file and the helpful links to examples.
We added the following `CONTRIBUTING.md` file: https://github.com/unfoldtoolbox/UnfoldSim.jl/blob/main/CONTRIBUTING.md
Please let us know whether this matches your ideas about the contribution file.
