### 2024-08-06 #6641 @jschepers
Dear @cbrnr,
thank you for your detailed suggestions about the paper. Please find our changes to the paper and additional explanations in the following PR: https://github.com/unfoldtoolbox/UnfoldSim.jl/pull/103
Please let us know if your questions have not yet been sufficiently answered or if additional changes should be made.
