### 2024-08-06 #6641 @jschepers
@cbrnr Thanks a lot for pointing this out. Luis Lips' contribution to the project is
indeed not clear from the UnfoldSim repository alone. However, in his master thesis (that @behinger and I supervised), he worked on "Statistically evaluating mixed-effects models for EEG analysis using large-scale simulations" and one part of his thesis was about implementing a framework to simulate EEG data. A revised version of this became the foundation for the UnfoldSim package.
His master thesis code can be found in [this repository](https://github.com/s-ccs/MSc_LMM_EEG).
