### 2024-08-06 #6641 @jschepers
Dear @roualdes,
thanks a lot for reviewing the package documentation. As developer of the package, one often becomes blind to many aspects, and your perspective has been very helpful to restructure the documentation and make it more understandable.
Please find our revision of the documentation in this PR: https://github.com/unfoldtoolbox/UnfoldSim.jl/pull/101
Please let us know if this addresses your suggestions regarding the package documentation or if anything is still missing
