### 2024-08-06 #6641 @jschepers
@aurorarossi and @cbrnr Thank you for your comments about the length of the paper. We can definitely see that our paper draft is rather long for a JOSS paper. However, we would like to keep both the simulation example and the references to other packages:

We think that the simulation example gives a good overview of the functionalities of the paper and offers readers a good starting point (even before having a look at the documentation). Additionally, it would be helpful to have a non-modifiable record of the tool's use cases for academic purposes (especially to use it in a cumulative dissertation). One alternative option would be to provide an additional arXiv version, but that seems cumbersome and confusing.

Regarding the package references, we would like to follow the idea of "giving credit where credit is due" and want to highlight the efforts that researchers put into research software development. This potentially helps establish Research Software Engineering as a viable academic path.

Given that the statement about the word count is not a "must" formulation we hope that our explanations convince you to keep the manuscript also with a quite extended length.
