### 2024-08-06 #6641 @cbrnr
Thanks for addressing my points, as I've already mentioned in https://github.com/unfoldtoolbox/UnfoldSim.jl/pull/103, I'm very happy with the changes! Regarding the paper length, I totally understand that these things are important, but given the journal constraints I'd still say there's room for reducing the length a bit. However, I'll leave this for the journal editors @mooniean @danielskatz to decide.

I'll be reviewing the software package next, but please note that I am currently busy with other things and will be on vacation soon, so it might take me until mid-September to finalize my review. I hope that's OK.
