### 2024-08-06 68c1786
Address joss paper comments (#103)
* Formatting

* Formatting

* adapted Unfold epoch function for Unfoldsim

* adapted simulate function

* added new tests for the dimensions of the simulate output

* Formatting

* Formatting

* adapted Unfold epoch function for Unfoldsim

* adapted simulate function

* added new tests for the dimensions of the simulate output

* Apply suggestions from code review (JuliaFormatter)

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Apply suggestions from code review (JuliaFormatter)

Co-authored-by: github-actions[b
[…]
