+++ test/design.jl
@@ -2,17 +2,19 @@
 
     @testset "SingleSubjectDesign" begin
         des = SingleSubjectDesign(; conditions = Dict(:A => nlevels(5), :B => nlevels(2)))
-        @test generate(des) == generate(des)
-        @test size(generate(des)) == (5 * 2, 2)
-        @test names(generate(des)) == ["A", "B"]
-        @test sum(generate(des).A .== "S2") == 2
-        @test sum(generate(des).B .== "S2") == 5
-        @test sum((generate(des).B .== "S2") .&& (generate(des).A .== "S2")) == 1
+        @test generate_events(des) == generate_events(des)
+        @test size(generate_events(des)) == (5 * 2, 2)
+        @test names(generate_events(des)) == ["A", "B"]
+        @test sum(generate_events(des).A .== "S2") == 2
+        @test sum(generate_events(des).B .== "S2") == 5
+        @test sum(
+            (generate_events(des).B .== "S2") .&& (generate_events(des).A .== "S2"),
+        ) == 1
         des = SingleSubjectDesign(;
             conditions = Dict(:A => nlevels(5), :B => nlevels(2)),
-            tableModifyFun = x -> sort(x, order(:B, rev = true)),
+            event_order_function = x -> sort(x, order(:B, rev = true)),
         )
-        @test generate(des).B[1] == "S2"
+        @test generate_events(des).B[1] == "S2"
     end
 
     @testset "MultiSubjectDesign" begin
@@ -23,28 +25,28 @@
         )
 
 
-        @test size(generate(des)) == (10 * 100 * 5 * 2, 4)
-        @test names(generate(des)) == ["subject", "item", "A", "B"]
-        @test sum(generate(des).A .== "S2") == 10 * 100 * 2
-        @test sum(generate(des).B .== "S2") == 10 * 100 * 5
+        @test size(generate_events(des)) == (10 * 100 * 5 * 2, 4)
+        @test names(generate_events(des)) == ["subject", "item", "A", "B"]
+        @test sum(generate_events(des).A .== "S2") == 10 * 100 * 2
+        @test sum(generate_events(des).B .== "S2") == 10 * 100 * 5
 
         # check that finally everything is sorted by subject
         des = MultiSubjectDesign(;
             n_subjects = 10,
             n_items = 100,
             both_within = Dict(:A => nlevels(5), :B => nlevels(2)),
-            tableModifyFun = x -> sort(x, order(:item, rev = true)),
+            event_order_function = x -> sort(x, order(:item, rev = true)),
         )
-        @test generate(des).subject[1] == "S01"
+        @test generate_events(des).subject[1] == "S01"
 
-        # check tableModifyFun
+        # check event_order_function
         des = MultiSubjectDesign(;
             n_subjects = 10,
             n_items = 100,
             both_within = Dict(:A => nlevels(5), :B => nlevels(2)),
-            tableModifyFun = x -> sort(x, order(:B, rev = true)),
+            event_order_function = x -> sort(x, order(:B, rev = true)),
         )
-        @test generate(des).B[1] == "S2"
+        @test generate_events(des).B[1] == "S2"
 
         # check that this throws an error because of `dv` as condition name
         des = MultiSubjectDesign(;
@@ -52,7 +54,7 @@
             n_items = 100,
             both_within = Dict(:dv => nlevels(5), :B => nlevels(2)),
         )
-        @test_throws AssertionError generate(des)
+        @test_throws AssertionError generate_events(des)
     end
 
     @testset "RepeatDesign" begin
@@ -65,7 +67,8 @@
         )
 
         design = RepeatDesign(designOnce, 3)
-        @test size(generate(design)) == (8 * 12 * 3, 3)
[…]

[…]
