### 2024-08-09 e05a9c6
merge fix, double definition of function
