+++ src/simulation.jl
@@ -7,10 +7,6 @@ Simulation(
 ) = Simulation(design, [component], onset, noisetype)
 
 
-function simulate(design::AbstractDesign, signal, onset::AbstractOnset, args...; kwargs...)
-    @warn "No random generator defined, used the default (`Random.MersenneTwister(1)`) with a fixed seed. This will always return the same results and the user is strongly encouraged to provide their own random generator!"
-    simulate(MersenneTwister(1), design, signal, onset, args...; kwargs...)
-end
 
 """
     simulate(
@@ -39,7 +35,6 @@ Some remarks to how the noise is added:
     If no `design.rng` was defined for `SequenceDesign`, we replace it with the `simulation`-function call `rng` object
 
 """
-
 
 function simulate(design::AbstractDesign, signal, onset::AbstractOnset, args...; kwargs...)
     @warn "No random generator defined, used the default (`Random.MersenneTwister(1)`) with a fixed seed. This will always return the same results and the user is strongly encouraged to provide their own random generator!"
