+++ docs/literate/HowTo/componentfunction.jl
@@ -36,4 +36,24 @@ erp = UnfoldSim.simulate_component(MersenneTwister(1), signal, design);
 
 
 # Finally, let's plot it, sorted by duration
-plot_erpimage(erp, sortvalues = generate_events(design).duration)
+
+f = Figure()
+df = UnfoldMakie.eeg_array_to_dataframe(erp')
+df.duration = repeat(generate_events(design).duration, inner = size(erp, 1))
+plot_erp!(
+    f[1, 1],
+    df,
+    mapping = (; group = :duration, color = :duration),
+    categorical_color = false,
+    categorical_group = true,
+    layout = (; legend_position = :left),
+)
+plot_erpimage!(
+    f[2, 1],
+    erp,
+    sortvalues = generate_events(design).duration,
+    layout = (; legend_position = :bottom),
+)
+f
+
+# The scaling by the two `condition`` effect levels and the modified event duration by the `duration` are clearly visible
