### 2024-08-09 #6641 @danielskatz
👋 sorry to have missed the notification above. Given that this paper is slightly over 2000 words, and the JOSS standard is up to 1000, I think some of it does need to be cut. My two suggestions to cut this and the overall number of pages to make it more like other JOSS papers are
1. Either change the functionality section to be a single paragraph, with the details in the documentation, or move the example to the documentation. I think the first would be better, as I like having an example in the paper.
2. Change the package references section to be a paragraph with items separated by commas, rather than having each item be its own line/paragraph.
