### 2024-08-21 #6641 @danielskatz
@jschepers - What is your plan for reducing the paper? And have you started on it?
