### 2024-08-22 #6641 @jschepers
@danielskatz I'm sorry for the late reply. I am currently on vacation/conferences and will be back in the office at the beginning of September. Thank you for your suggestions regarding the paper. I will implement the changes to the package references section and will discuss the shortening options with Benedikt. We will also go through the whole paper again and decide which information we can leave out. Since the word limit is not a "must" formulation we hoped that we could keep the paper in its current structure because we think that the extra information is valuable to the readers. Moreover, it would be valuable to me to have the paper in its long form for a cumulative dissertation. As discussed before, we have to check back with the University committee if we need a "full" ArXiv version, which would be a bit redundant. But of course we also understand the goal of JOSS to keep papers within the word limit.
If we do not find enough other content to shorten I currently have a tendency to changing the functionality section to a single paragraph because I also would like to keep the example.
