### 2024-09-03 612786a
fix tutorial with v0.3 renaming of simulate to simulate_component
