### 2024-09-03 272eb02
fix newComponent tutorial
