### 2024-09-20 unfoldtoolbox/UnfoldSim.jl#106 @cbrnr
Follow docstring conventions
This issue is part of https://github.com/openjournals/joss-reviews/issues/6641.

As @roualdes has already mentioned in #95, the docstring for `predef_eeg` seems to be broken. Since this is likely one of the first functions that new users will interact with, I think it is really important to get this docstring right. I'm not sure what's going on, but it seems like underscores are mangled with the following characters and interpreted as an underline.

![Screenshot 2024-09-20 at 08 31 46](https://github.com/user-attachments/assets/c1a26521-3288-4b03-bcc8-8dc570991ccf)

In addition, I strongly suggest to follow Julia standard docstring conventions. Currently, the structure differs quite a lot from what I'd expect. For example, I find it quite difficult to get to know the signature when the various kwargs are not listed in a single place. This is especially important because usually I'd like to see all the default values just to know which arguments are required.

If you compare this to the doc of e.g. `LinearAlgebra.svd`, I'd recommend to implement something that follows what they are doing (which seems to be the standard):

![Screenshot 2024-09-20 at 08 36 59](https://github.com/user-attachments/assets/7c02f0ea-cdf0-43f1-9401-22539826c748)
