### 2024-09-20 unfoldtoolbox/UnfoldSim.jl#106 @behinger
hi!
Thanks for the suggestion; we will think about the style of docs - can you check the branch #joss-doc-suggestion ? We fixed quite a bit of docstrings already (e.g. your raised point about the `_` should be fixed
