### 2024-09-20 unfoldtoolbox/UnfoldSim.jl#106 @cbrnr
Hm, I installed this branch, but the docstring doesn't look any different, not sure if I did it correctly?

```
(@v1.10) pkg> st
Status `~/.julia/environments/v1.10/Project.toml`
  [ed8ae6d2] UnfoldSim v0.3.2 `https://github.com/unfoldtoolbox/UnfoldSim.jl#joss-doc-suggestions`
```

I also wanted to add that my comment about default arguments applies to other functions/objects as well, e.g. `SingleSubjectDesign` does not mention which arguments are required and which are optional (and their defaults). I assume that this will also be the case for other functions in this package, and this makes it rather difficult to learn the API. What I would like to see, and I think this is also standard in Julia, is all arguments (and their defaults for optional args) at a glance.
