### 2024-09-20 unfoldtoolbox/UnfoldSim.jl#106 @behinger
![grafik](https://github.com/user-attachments/assets/a521e8b3-2692-4024-b745-0d0390163d4a)
strange. Not sure what's going on?

Your point is well taken, and we will try to work on all docstrings but it will take a bit of time
