### 2024-09-20 unfoldtoolbox/UnfoldSim.jl#106 @cbrnr
> strange. Not sure what's going on?

No idea either. I simply installed it with

```
]
add UnfoldSim#joss-doc-suggestions
```

This should be correct?
