### 2024-09-20 unfoldtoolbox/UnfoldSim.jl#106 @behinger
Did you restart the REPL? Sometimes I have noticed it doesnt update the package - I just tried with `]activate -temp` and installed as you mentioned.

can you doublecheck with `pathof(UnfoldSim)` that it points to `UnfoldSim/lM1mg/src/UnfoldSim.jl` this hash (iirc the hash should be the same regardless of julia version)?
