### 2024-09-20 #6641 @cbrnr
I have an additional comment regarding the paper. In the "Simulation example" section, some of the code uses functions from other packages, e.g.

```julia
design = SingleSubjectDesign(;
    conditions = Dict(
        :condition => ["car", "face"],
        :continuous => range(0, 5, length = 10),
    ),
    event_order_function = x -> shuffle(deepcopy(StableRNG(1)), x),
) |> x -> RepeatDesign(x, 100)
```

uses StableRNG from `StableRNGs.jl` and shuffle from `Random`. I think for the sake of reproducibility, the `using ...` lines for standard library modules should be explicitly added (maybe at the beginning of the section). Regarding third-party packages, I think it would be nice to try to avoid them as much as possible here.

For example, and for the sake of simplicity, would it not be sufficient to use

```julia
event_order_function = x -> shuffle(MersenneTwister(1), x),
```

here (provided that `using random` has been run)?

Bonus question just for me: What is the purpose of `deepcopy` here? If you want to use the RNG in subsequent lines, why not initialize it with

```julia
rng = MersenneTwister(1)
```

and then you can use `rng` whenever you need it?
