### 2024-09-20 #6641 @behinger
- we use `StableRNG` because imho it is much more accessible function name than `MersenneTwister`
- the deepcopy is a relic from another place and after some simple testing, can be removed. But we cannot use `rng = MersenneTwister(1)` because internally we might call `generate_events(SinglesubjectDesign)` multiple times and need the exact same output, which with a single rng object cannot be made. So either
```julia
rng = MersenneTwister(1)
...
event_order_function = x->shuffle(deepcopy(rng),x)
```
or
```julia
event_order_function = x->shuffle(MersenneTwister(1),x)
```
if there is an easier way, happy to do that as well - but thanks for the question, the deepcopy is really unnecessary here.
