### 2024-09-20 unfoldtoolbox/UnfoldSim.jl#106 @cbrnr
> Did you restart the REPL? Sometimes I have noticed it doesnt update the package - I just tried with `]activate -temp` and installed as you mentioned.

Ah yes, restarting the REPL fixed it! Thanks! The comment on following established standards (specifically with respect to arguments) still stands though.
