### 2024-09-20 #6641 @cbrnr
Right, but for the sake of showing a simple example, I'd prefer the second option here (even though I agree the name `MersenneTwister` is not as self-explanatory). The goal here should be to reduce complexity and focus on your package.

Regarding the question on getting the exact same numbers, that's a design choice. If you have already implemented it in a specific way, I don't want to ask for any changes, but it would be nice if the internal implementation was independent from any RNG state passed in (and used) in user code.

For this particular example, I think it doesn't matter if the order is reproducible or not, so you could also just write

```julia
using Random

design = SingleSubjectDesign(;
    conditions = Dict(
        :condition => ["car", "face"],
        :continuous => range(0, 5, length = 10),
    ),
    event_order_function = x -> shuffle(x),
) |> x -> RepeatDesign(x, 100)
```
