+++ src/component.jl
@@ -285,6 +285,34 @@ function weight_σs(σs::Dict, b_σs::Float64, σ_lmm::Float64)
     return named_random_effects
 end
 
+
+#--- TRF Component
+
+"""
+    TRFComponent(components::Vector{<:AbstractComponents},features::AbstractArray)
+
+    This component can be used to convolve a `response` of a component-vector with a feature-array.
+
+Each column of the feature-array will be convolved with one generated response from the AbstractComponent-vector (that is, each row of the respective `AbstractDesign`)
+
+If only a single TRF is needed, a vector can be provided.
+"""
+struct TRFComponent <: AbstractComponent
+    components
+    features::AbstractArray
+end
+UnfoldSim.length(t::TRFComponent) = size(t.features,1)
+
+
+function UnfoldSim.simulate_component(rng,c::TRFComponent,design::AbstractDesign)
+    kernel = UnfoldSim.simulate_responses(rng,c.components,UnfoldSim.Simulation(design,c,NoOnset(),NoNoise()))
+
+    @assert size(kernel,2) == size(c.features,2) "if the $(typeof(design)) generates multiple columns (sz=$(size(kernel))), the $(typeof(c)) needs to have multiple columns as well (sz=$(size(c.signal)))"
+    x = reduce(hcat,UnfoldSim.conv.(eachcol(c.features),eachcol(kernel)))[1:size(c.features,1),:]
+    return x
+
+end
+
 #----
 
 """
