+++ src/onset.jl
@@ -42,7 +42,7 @@ In the case that a TRF is simulated (via `TRFComponent`). This onset returns a v
 """
 struct TRFOnset <: AbstractOnset end
 
-function UnfoldSim.simulate_interonset_distances(rng,onset::TRFOnset,design)
+function simulate_interonset_distances(rng,onset::TRFOnset,design)
 	sz = size(design)
 	return Int.(zeros(sz))
 end
@@ -150,10 +150,10 @@ end
 function simulate_interonset_distances(rng, o::UniformOnsetFormula, design::AbstractDesign)
     events = generate_events(design)
     widths =
-        UnfoldSim.generate_designmatrix(o.width_formula, events, o.width_contrasts) *
+        generate_designmatrix(o.width_formula, events, o.width_contrasts) *
         o.width_β
     offsets =
-        UnfoldSim.generate_designmatrix(o.offset_formula, events, o.offset_contrasts) *
+        generate_designmatrix(o.offset_formula, events, o.offset_contrasts) *
         o.offset_β
 
     return Int.(
@@ -208,10 +208,10 @@ function simulate_interonset_distances(
     events = generate_events(design)
 
 
-    μs = UnfoldSim.generate_designmatrix(o.μ_formula, events, o.μ_contrasts) * o.μ_β
-    σs = UnfoldSim.generate_designmatrix(o.σ_formula, events, o.σ_contrasts) * o.σ_β
+    μs = generate_designmatrix(o.μ_formula, events, o.μ_contrasts) * o.μ_β
+    σs = generate_designmatrix(o.σ_formula, events, o.σ_contrasts) * o.σ_β
     offsets =
-        UnfoldSim.generate_designmatrix(o.offset_formula, events, o.offset_contrasts) *
+        generate_designmatrix(o.offset_formula, events, o.offset_contrasts) *
         o.offset_β
 
 
