+++ previews/PR107/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"EditURL = \"../../../literate/HowTo/sequence.jl\"","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"using Base: add_sum\nusing UnfoldSim\nusing CairoMakie\nusing StableRNGs","category":"page"},{"location":"generated/HowTo/sequence/#Stimulus-Response-design","page":"Produce specific sequences of events","title":"Stimulus - Response design","text":"","category":"section"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"Let's say we want to simulate a stimulus response, followed by a button press response.","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"First we generate the minimal design of the experiment by specifying our conditins (a one-condition-two-levels design in our case)","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"design = SingleSubjectDesign(conditions = Dict(:condition => [\"one\", \"two\"]))\ngenerate_events(design)","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"Next we use the SequenceDesign and nest our initial design in it. \"SR_\" is code for an \"S\" event and an \"R\" event - only single letter events are supported! The \"_\" is a signal for the Onset generator to generate a bigger pause - no overlap between adjacend \"SR\" pairs","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"design = SequenceDesign(design, \"SR_\", StableRNG(1))\ngenerate_events(design)","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"The main thing that happened is that the design was repeated for every event (each 'letter') of the sequence, and an eventtype column was added.","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"hint: Hint\nmore advaned sequences are possible as well, like \"SR{1,3}\", or \"A[BC]\". Infinite sequences are not possible like \"AB*\".","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"Finally, let's repeat the current design 4 times","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"design = RepeatDesign(design, 4)\ngenerate_events(design)","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"This results in 16 trials that nicely follow our sequence","category":"page"},{"location":"generated/HowTo/sequence/","page":"Produce specific sequences of events","title":"Produce specific sequences of events","text":"hint: Hint\nThere is a difference between SequenceDesign(RepeatDesign) and RepeatDesign(SequenceDesign) for variable sequences e.g. \"A[BC]\", where in the former case,  one sequence is drawn e.g. \"AC\" and applied to all repeated rows, in the latter, one sequence for each repeat is drawn.","category":"page"},{"location":"generated
[…]
