### 2024-09-27 966706a
Adapted package reference section
