### 2024-09-27 b30bb3c
Test whether figure creation gh action works
