+++ joss_paper/create_paper_plots.jl
@@ -317,3 +317,5 @@ let
     current_figure()
 end
 #----
+
+#Test
