### 2024-09-27 88e2e7a
start shortening paper
