+++ joss_paper/create_paper_plots.jl
@@ -317,5 +317,3 @@ let
     current_figure()
 end
 #----
-
-#Test
