### 2024-09-27 38c3a88
start shortening paper
