### 2024-09-30 a4c4131
Continue shortening paper
