### 2024-09-30 d256504
Continue shortening paper
