### 2024-10-09 f7122ea
Continue shortening paper
