### 2024-10-09 250e23a
Continue shortening paper
