### 2024-10-09 3c5735d
Continue shortening paper
