### 2024-10-09 c1d26c4
Continue shortening paper
