### 2024-10-09 d693d90
Continue shortening paper
