### 2024-10-09 8f0f164
Add references for programming languages
