+++ joss_paper/paper.bib
@@ -294,6 +294,13 @@ url = {https://doi.org/10.5281/zenodo.8344531}
 
   % MATLAB
 
+@software{MATLAB,
+  author = {The MathWorks Inc.},
+  publisher = {The MathWorks Inc.},
+  address = {Natick, Massachusetts, United States},
+  url = {https://www.mathworks.com}
+}
+
 @article{vaziri2023eegg,
   title={EEGg Generating Synthetic EEG signals in Matlab environment},
   author={Vaziri, Ava Yektaeian and Makkiabadi, Bahador and Samadzadehaghdam, Nasser},
@@ -385,4 +392,14 @@ url = {https://doi.org/10.5281/zenodo.8344531}
   year={2018},
   publisher={IEEE},
   doi={10.1109/TNSRE.2018.2873061}
+}
+
+% R
+@Manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2021},
+  url = {https://www.R-project.org/},
 }
