### 2024-10-09 8c682ad
Add references for programming languages
