### 2024-10-09 e992dec
Continue shortening paper
