### 2024-10-09 d983a2b
Continue shortening paper
