+++ joss_paper/paper.bib
@@ -386,13 +386,3 @@ url = {https://doi.org/10.5281/zenodo.8344531}
   publisher={IEEE},
   doi={10.1109/TNSRE.2018.2873061}
 }
-
-% R
-@Manual{R,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2021},
-  url = {https://www.R-project.org/},
-}
