### 2024-10-10 ba3e915
Continue shortening paper
