### 2024-10-10 17f7477
Adapted date
