### 2024-10-10 10b2baf
adapt code snippet
