### 2024-10-10 #6641 @behinger
Dear all,
we finished shortening the paper. It's shortened by ~200 words (~15%) and the content spans one page less than before. We really tried, but without removing additional content, we don't think it is possible. Given that the 1000 Word rule is a "can" not a "must" rule, we hope that this is satisfactory. If not, we really have to cut deep and remove more - we do like it like it is now though, maybe an explanation of the rational would be helpful for us to better understand why a shorter paper is preferred.

Moreover, we adapted the code snippets according to @cbrnr 's suggestions, i.e. we added package imports at the beginning, changed the RNG from `StableRNG` to `MersenneTwister`; we also came up with a way to have only a single user-facing random generator, thus in future it will be possible to do `event_order_function = shuffle` directly, by requiring a two argument function e.g. `(rng::AbstractRNG,x)->my_shuffle(rng,x)`. Thanks for suggesting an improvement here - we still need to implement it, but it is straight forward; except semester starts and we are scrambling for time right now.

Best, Bene and Judith
