### 2024-10-10 #6641 @danielskatz
checking on current word count...  And I agree that 1000 words is a goal, not a requirement.  And I defer to @mooniean as editor to decide on next steps.
