### 2024-10-11 #6641 @mooniean
Hi @jschepers and @behinger! Thank you for your revision. However, the paper is still close to twice the recommended size. I would suggest, for example, to leave the simulation section in the documentation and keep one line referencing it, as currently is on line 64.
