+++ .github/workflows/CI.yml
@@ -3,7 +3,7 @@ on:
   push:
     branches:
       - main
-    tags: ["*"]
+    tags: '*'
   pull_request:
     branches:
       - main
