+++ docs/make.jl
@@ -58,7 +58,7 @@ makedocs(;
 
 deploydocs(;
     repo = "github.com/unfoldtoolbox/UnfoldSim.jl",
-    devbranch = "main",
-    versions = "v#.#",
+    #devbranch = "main",
+    #versions = "v#.#",
     push_preview = true,
 )
