+++ v0.3.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"EditURL = \"../../../literate/HowTo/newComponent.jl\"","category":"page"},{"location":"generated/HowTo/newComponent/#Define-a-new-component-(with-variable-duration-and-shift)","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"","category":"section"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"We want a new component that changes its duration and shift depending on a column in the event design. This is somewhat already implemented in the HRF + Pupil bases.","category":"page"},{"location":"generated/HowTo/newComponent/#Setup","page":"Define a new component (with variable duration and shift)","title":"Setup","text":"","category":"section"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"<details>\n<summary>Click to expand</summary>","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"using UnfoldSim\nusing Unfold\nusing Random\nusing DSP\nusing CairoMakie, UnfoldMakie\n\nsfreq = 100;\nnothing #hide","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"</details >","category":"page"},{"location":"generated/HowTo/newComponent/#Design","page":"Define a new component (with variable duration and shift)","title":"Design","text":"","category":"section"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"Let's generate a design with two columns, shift + duration","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"design = UnfoldSim.SingleSubjectDesign(;\n    conditions = Dict(\n        :shift => rand(100) .* sfreq / 5,\n        :duration => 20 .+ rand(100) .* sfreq / 5,\n    ),\n)","category":"page"},{"location":"generated/HowTo/newComponent/#Implement-a-new-AbstractComponent","page":"Define a new component (with variable duration and shift)","title":"Implement a new AbstractComponent","text":"","category":"section"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"We also need a new AbstractComponent","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"struct TimeVaryingComponent <: AbstractComponent\n    basisfunction::Any\n    maxlength::Any\nend","category":"page"},{"location":"generated/HowTo/newComponent/","page":"Define a new component (with variable duration and shift)","title":"Define a new component (with variable duration and shift)","text":"We have to define the length of a component","category":"page"},{"location":"generated/HowTo/newComponent/","page
[…]
