+++ versions.js
@@ -1,2 +1,7 @@
 var DOC_VERSIONS = [
+  "stable",
+  "v0.3",
+  "dev",
 ];
+var DOCUMENTER_NEWEST = "v0.3.2";
+var DOCUMENTER_STABLE = "stable";
