### 2024-10-16 360b0eb
add GroundTruth desing and needed functions
