### 2024-10-16 876fc69
Start adapting simulate ERP tutorial
