### 2024-10-16 199ae7c
Update make.jl
Delete commented code which is not needed
