### 2024-10-16 fea3f2d
Update README.md
Add badge with link to the stable documentation
