### 2024-10-16 bcc68e6
Update README.md
Link to stable version of the documentation.
