### 2024-10-22 85b8221
Added the content from the paper simulation example to the docs example
