+++ docs/literate/tutorials/simulateERP.jl
@@ -28,7 +28,9 @@ using UnfoldMakie # For plotting
 # ```
 
 # ## Specify the simulation "ingredients"
-# 1\. We specify an **experimental design** with one subject in two experimental conditions including a continuous variable with 10 values.
+
+# ### 1. Experimental design
+# We specify an experimental design with one subject in two experimental conditions including a continuous variable with 10 values.
 # To mimic randomization in an experiment, we shuffle the trials using the `event_order_function` argument. To generate more trials we repeat the design 100 times which results in 2000 trials in total.
 design =
     SingleSubjectDesign(;
@@ -45,76 +47,169 @@ first(events_df, 5)
 # Above you can see the first five rows extracted from the events data frame representing the experimental design. Each row corresponds to one event.
 # The columns *continuous* and *condition* display the levels of the predictor variables for the specific event.
 
-# 2\. Next, we create a signal consisting of two different **components**.
-# For the first component, we use the prespecified N170 base with an intercept of 5µV and a condition effect of 3&nbsp;µV for the “face/car” condition i.e. faces will have a more negative signal than cars.
-# For the second component, we use the prespecified P300 base and include a linear and a quadratic effect of the continuous variable: the larger the value of the continuous variable, the larger the simulated potential.
+# ### 2. Event basis functions (Components)
+# Next, we create a signal consisting of three different components.
 
-# **p100** is unaffected by our design and has amplitude of 5
+# For the first component, we use the prespecified **P100** base which will be unaffected by our design and has an amplitude of 5 µV.
 p1 = LinearModelComponent(; basis = p100(), formula = @formula(0 ~ 1), β = [5]);
 
-# **n170** has a condition effect, faces are more negative than cars
+# For the second component, we use the prespecified **N170** base with an intercept of 5 µV and a condition effect of 3 µV for the “face/car” condition i.e. faces will have a more negative signal than cars.
 n1 = LinearModelComponent(;
     basis = n170(),
     formula = @formula(0 ~ 1 + condition),
     β = [5, 3],
 );
-# **p300** has a continuous effect, higher continuous values will result in larger P300's.
-# We include both a linear and a quadratic effect of the continuous variable.
+# For the third component, we use the prespecified **P300** base and include a linear and a quadratic effect of the continuous variable: the larger the value of the continuous variable, the larger the simulated potential.
 p3 = LinearModelComponent(;
     basis = p300(),
     formula = @formula(0 ~ 1 + continuous + continuous^2),
     β = [5, 1, 0.2],
 );
 
-# Now we can simply combine the components and simulate 
+# ### 3. Inter-onset distribution
+# In the next step, we specify an inter-onset distribution, in this case, a uniform distribution with an inter-event distance of exactly 200 samples.
+onset = UniformOnset(; width = 0, offset = 200);
+
+# ### 4. Noise specification
+# As the last ingredient, we specify the **noise**, in this case, Pink noise.
+noise = PinkNoise(; noiselevel = 2);
+
+# ## Simulate data
+
+# Finally, we combine all the ingredients and simulate data. To make the simulation reproducible, one can specify a random generator.
+
+## Combine the components in a vector
 components = [p1, n1, p3]
-data, evts = simulate(
-    MersenneTwister(1),
-    design,
-    components,
-    UniformOnset(; width = 0, offset = 1000),
[…]
