### 2024-10-28 f082ce4
implement EffectsDesign; implement Tests; export Design
