### 2024-10-28 ba10e1d
fix size(EffectsDesign) bug
