+++ src/design.jl
@@ -208,7 +208,7 @@ struct EffectsDesign <: AbstractDesign
     effects_dict::Dict
 end
 EffectsDesign(design::MultiSubjectDesign,effects_dict::Dict) = error("not yet implemented")
-UnfoldSim.size(t::EffectsDesign) = size(generate_events(t))
+UnfoldSim.size(t::EffectsDesign) = size(generate_events(t),1)
 
 """
     expand_grid(design)
