### 2024-10-28 f5facfe
fix bug in EffectsDesign test
