### 2024-10-29 a10d5a9
Update getGroundTruth.jl tutorial
