### 2024-11-04 b7df987
unified interface to only use one rng
