+++ test/sequence.jl
@@ -15,7 +15,7 @@ end
 
 
     design = SingleSubjectDesign(conditions = Dict(:condition => ["one", "two"]))
-    design = SequenceDesign(design, "SCR_", StableRNG(1))
+    design = SequenceDesign(design, "SCR_")
     evt = generate_events(design)
     @test size(evt, 1) == 6
     @test evt.event == ['S', 'C', 'R', 'S', 'C', 'R']
@@ -29,16 +29,16 @@ end
     # repeat first, then sequence => same sequence
     design = SingleSubjectDesign(conditions = Dict(:condition => ["A", "B"]))
     design = RepeatDesign(design, 2)
-    design = SequenceDesign(design, "S[ABCD]", StableRNG(2))
-    evt = generate_events(design)
+    design = SequenceDesign(design, "S[ABCD]")
+    evt = generate_events(StableRNG(2), design)
 
     @test all(evt.event[2:2:end] .== 'B')
 
 
     # sequence first, then repeat => different sequence for each repetition
     design = SingleSubjectDesign(conditions = Dict(:condition => ["A", "B"]))
-    design = SequenceDesign(design, "S[ABCD]", StableRNG(2))
+    design = SequenceDesign(design, "S[ABCD]")
     design = RepeatDesign(design, 2)
-    evt = generate_events(design)
+    evt = generate_events(StableRNG(2), design)
     @test !all(evt.event[2:2:end] .== 'B')
 end
