+++ src/design.jl
@@ -9,7 +9,7 @@ A type for specifying the experimental design for multiple subjects (based on th
 - `subjects_between` = Dict{Symbol,Vector} -> effects between subjects, e.g. young vs old 
 - `items_between` = Dict{Symbol,Vector} -> effects between items, e.g. natural vs artificial images, (but shown to all subjects if not specified also in `subjects_between`)
 - `both_within` = Dict{Symbol,Vector}	-> effects completly crossed
-- `event_order_function` = `x->x`; # can be used to sort, or e.g. `x->shuffle(MersenneTwister(42),x)` - be sure to fix/update the rng accordingly!!
+- `event_order_function` = `(rng,x)->x`; # can be used to sort, or e.g. `(rng,x)->shuffle(rng,x)` (or shorter just `event_order_function=shuffle`)
 
 Tip: Check the resulting dataframe using `generate_events(design)`
 
@@ -32,7 +32,7 @@ See also [`SingleSubjectDesign`](@ref), [`RepeatDesign`](@ref)
     subjects_between::Dict{Symbol,Vector} = Dict()
     items_between::Dict{Symbol,Vector} = Dict()
     both_within::Dict{Symbol,Vector} = Dict()
-    event_order_function = x -> x # can be used to sort, or x->shuffle(rng,x)
+    event_order_function = (rng, x) -> x
 end
 
 
@@ -42,8 +42,8 @@ end
 A type for specifying the experimental for a single subject (based on the given conditions).
 
 ### Fields
-- conditions = Dict{Symbol,Vector} of conditions, e.g. `Dict(:A=>["a_small","a_big"],:B=>["b_tiny","b_large"])`
-- `event_order_function` = x->x; # can be used to sort, or x->shuffle(MersenneTwister(42),x) - be sure to fix/update the rng accordingly!!
+- `conditions = Dict{Symbol,Vector}`` of conditions, e.g. `Dict(:A=>["a_small","a_big"],:B=>["b_tiny","b_large"])`
+- `event_order_function` = (rng::AbstractRNG,x::DataFrame)->x; # can be used to sort by specifying `sort`, or shuffling by providing `shuffle`, or custom functions following the interface `(rng,x)->my_shuffle(rng,x)`
 
 Number of trials / rows in `generate_events(design)` depend on the full factorial of your `conditions`.
 
@@ -65,7 +65,7 @@ See also [`MultiSubjectDesign`](@ref), [`RepeatDesign`](@ref)
 """
 @with_kw struct SingleSubjectDesign <: AbstractDesign
     conditions::Dict{Symbol,Vector} = Dict()
-    event_order_function = x -> x
+    event_order_function = (rng, x) -> x
 end
 
 
@@ -85,7 +85,7 @@ If conditions is `nothing`, a single trial is simulated with a column `:dummy` a
 julia> d = SingleSubjectDesign(;conditions= Dict(:A=>nlevels(5),:B=>nlevels(2)))
 julia> generate_events(d)
 """
-function generate_events(design::SingleSubjectDesign)
+function generate_events(rng::AbstractRNG, design::SingleSubjectDesign)
     if isempty(design.conditions)
         events = DataFrame(:dummy => [:dummy])
     else
@@ -96,24 +96,36 @@ function generate_events(design::SingleSubjectDesign)
             DataFrame
     end
     # by default does nothing
-    return design.event_order_function(events)
+    return apply_event_order_function(design.event_order_function, rng, events)
+
 end
 
+function apply_event_order_function(fun, rng, events)
+    try
+        return fun(rng, events)
+
+    catch e
+        error(
+            "Problem in `event_order_function` - Make sure the function allows for two inputs: `(rng::AbstractRNG,x::DataFrame)`",
+        )
+    end
+end
 """
[…]
