### 2024-11-04 598f8c2
fix @jscheper code review issues
