### 2024-11-04 3615a97
added tests for rng fix in event_sort function
