+++ test/design.jl
@@ -15,6 +15,17 @@
             event_order_function = (rng, x) -> sort(x, order(:B, rev = true)),
         )
         @test generate_events(des).B[1] == "S2"
+
+        des = SingleSubjectDesign(;
+            conditions = Dict(:A => nlevels(5), :B => nlevels(2)),
+            event_order_function = (rng, x) -> shuffle(rng, x),
+        )
+        @test generate_events(MersenneTwister(3), des) ==
+              generate_events(MersenneTwister(3), des)
+
+        # different sortig seed results in different sorts
+        @test generate_events(MersenneTwister(3), des) !=
+              generate_events(MersenneTwister(4), des)
     end
 
     @testset "MultiSubjectDesign" begin
@@ -47,6 +58,21 @@
             event_order_function = (rng, x) -> sort(x, order(:B, rev = true)),
         )
         @test generate_events(des).B[1] == "S2"
+
+        # check event_order_function
+        des = MultiSubjectDesign(;
+            n_subjects = 10,
+            n_items = 100,
+            both_within = Dict(:A => nlevels(5), :B => nlevels(2)),
+            event_order_function = (rng, x) -> shuffle(rng, x),
+        )
+        # generating same events with same seed should result in same events
+        @test generate_events(MersenneTwister(3), des) ==
+              generate_events(MersenneTwister(3), des)
+
+        # different sortig seed results in different sorts
+        @test generate_events(MersenneTwister(3), des) !=
+              generate_events(MersenneTwister(4), des)
 
         # check that this throws an error because of `dv` as condition name
         des = MultiSubjectDesign(;
@@ -163,3 +189,4 @@
     end
 
 end
+
