### 2024-11-04 975ebe4
fixed rng for event_order function
