### 2024-11-05 47ccb9e
Adapted imbalanced design tutorial
