### 2024-11-05 5a8dd61
JuliaFormatter: Apply suggestions from code review
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
