### 2024-11-05 c4f8bda
adapt code to event_order_function changes
