### 2024-11-05 e6e3401
Update CI.yml
Replaced Julia 1.8 with 1
