+++ .github/workflows/CI.yml
@@ -20,7 +20,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - "1.8"
+          - "1"
         os:
           - ubuntu-latest
         arch:
