### 2024-11-05 c1fd2ac
Removed simulation example section
