### 2024-11-06 830707f
Added missing rng or deepcopy(rng)
