### 2024-11-06 745cae4
add RNG to generate_events call
