### 2024-11-06 aa89c2c
JuliaFormatter: Apply suggestions from code review
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
