+++ docs/literate/HowTo/multichannel.jl
@@ -46,13 +46,9 @@ onset = UniformOnset(; width = 20, offset = 4);
 # ## Simulation
 
 # Now as usual we simulate data. Inspecting data shows our result is now indeed ~230 Electrodes large! Nice!
-data, events = simulate(
-    MersenneTwister(1),
-    design,
-    [mc, mc2],
-    onset,
-    PinkNoise(noiselevel = 0.05));
-    
+data, events =
+    simulate(MersenneTwister(1), design, [mc, mc2], onset, PinkNoise(noiselevel = 0.05));
+
 size(data)
 
 
