### 2024-11-07 93944df
Started formatting docstrings in components.jl
