### 2024-11-11 10f1d64
Formatted docstrings in bases.jl
