+++ src/bases.jl
@@ -1,46 +1,198 @@
-# here we define some commonly used basis for simulation
+# Here we define some commonly used basis functions for simulation
 
 ## EEG
 """
-    p100(;sfreq=100)
-Generator for Hanning window, peak at 100ms, width 100ms, at kwargs `sfreq` (default 100). Returns a vector.
+    p100(; sfreq = 100)
+
+Generate a Hanning window mimicking a P100 EEG component with a peak at 100ms and a width of 100ms.
+
+# Keyword arguments
+- `sfreq = 100`: Sampling frequency in Hz.
+
+# Returns
+- `Vector`: Contains a shifted (i.e. zero-padded) hanning window.
+
+# Examples
+```julia-repl
+julia> p100(; sfreq = 200)
+30-element Vector{Float64}:
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ ⋮
+ 0.37725725642960045
+ 0.22652592093878665
+ 0.10542974530180327
+ 0.02709137914968268
+ 0.0
+```
+
+See also [`p300`](@ref), [`n170`](@ref), [`n400`](@ref), [`hanning`](@ref).
 """
 p100(; sfreq = 100) = hanning(0.1, 0.1, sfreq)
 
 """
-    p300(;sfreq=100)
-Generator for Hanning window, peak at 300ms, width 300ms, at kwargs `sfreq` (default 100). Returns a vector.
+    p300(; sfreq = 100)
+
+Generate a Hanning window mimicking a P300 EEG component with a peak at 300ms and a width of 300ms.
+
+# Keyword arguments
+- `sfreq = 100`: Sampling frequency in Hz.
+
+# Returns
+- `Vector`: Contains a shifted (i.e. zero-padded) hanning window.
+
+# Examples
+```julia-repl
+julia> p300(; sfreq = 150)
+67-element Vector{Float64}:
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ ⋮
+ 0.07937323358440934
+ 0.04518400232274078
+ 0.02025351319275137
+ 0.005089279059533658
+ 0.0
+```
+
+See also [`p100`](@ref), [`n170`](@ref), [`n400`](@ref), [`hanning`](@ref).
 """
 p300(; sfreq = 100) = hanning(0.3, 0.3, sfreq)
 
 """
-    n170(;sfreq=100)
-Generator for Hanning window, negative (!) peak at 170ms, width 150ms, at kwargs `sfreq` (default 100). Returns a vector.
+    n170(; sfreq = 100)
+
+Generate a Hanning window mimicking an N170 EEG component with a negative (!) peak at 170ms and a width of 150ms.
+
+# Keyword arguments
+- `sfreq = 100`: Sampling frequency in Hz.
+
+# Returns
+- `Vector`: Contains a shifted (i.e. zero-padded) hanning window.
+
+# Examples
+```julia-repl
+julia> n170(; sfreq = 120)
+28-element Vector{Float64}:
+ -0.0
+ -0.0
+ -0.0
+ -0.0
+ -0.0
+  ⋮
+ -0.45386582026834904
+ -0.2771308221117309
+ -0.1304955413896705
+ -0.03376388529782215
+ -0.0
+```
+See also [`p100`](@ref), [`p300`](@ref), [`n400`](@ref), [`hanning`](@ref).
 """
 n170(; sfreq = 100) = -hanning(0.15, 0.17, sfreq)
 
 """
-    n400(;sfreq=100)
[…]
