### 2024-11-11 #6641 @jschepers
Hi @mooniean, Thanks a lot for checking in and for your patience. As you suggested, I removed the simulation section and only linked to an example in the documentation.

Additionally, as suggested by @cbrnr,  @maanikmarathe and I started to adapt the docstrings according to a template, based on the Julia documentation, a style guide and examples from other packages, to make them more consistent.
For details please see: https://github.com/unfoldtoolbox/UnfoldSim.jl/pull/121

Best,
Judith
