### 2024-11-11 a3849c6
better docstring PuRF and HRF
