### 2024-11-11 d0b51c6
return type HRF
