### 2024-11-20 #6641 @mooniean
Hi @jschepers! It's looking great and we're almost there!

I've read the latest version and have a few formatting changes:

- Line 29: No reference for MATLAB
- Line 32, 62: UnfoldSim.jl is not with the same formatting as the abstract
- Line 55: Not syre if <: is intended to show on the text
- Figure 1 Legend: "... while the distribution is shown on a linear scale" should perhaps be "... while the uniform distribution is shown on a linear scale"

In the comment above, you can see there's additional author tasks, such as archiving on zenodo and posting the DOI here. Please let me know when you've managed to implement the changes and have done the steps above!
