### 2024-11-20 #6641 @cbrnr
I'm not sure if I missed a deadline, but I haven't actually reviewed the software yet (other than my comment on docstring style). Should I still do that?
