### 2024-11-22 0d1aa87
format docstring in simulation.jl
