### 2024-11-25 4c8c5d2
Finished formatting the docstrings in component.jl
