### 2024-11-26 43dac8c
improved docstrings #108
