+++ src/component.jl
@@ -6,7 +6,7 @@ A component that adds a hierarchical relation between parameters according to a 
 All fields can be named. Works best with [`MultiSubjectDesign`](@ref).
 
 # Fields
-- `basis::Any`: an object, if accessed, provides a 'basis function', e.g. `hanning(40)`, this defines the response at a single event. It will be weighted by the model prediction.
+- `basis::Any`: an object, if accessed, provides a 'basis function', e.g. `hanning(40)::Vector`, this defines the response at a single event. It will be weighted by the model prediction. Future versions will allow for functions, as of v0.3 this is restricted to array-like objects
 - `formula::Any`: Formula-object in the style of MixedModels.jl e.g. `@formula 0 ~ 1 + cond + (1|subject)`. The left-hand side is ignored.
 - `β::Vector` Vector of betas (fixed effects), must fit the formula.
 - `σs::Dict` Dict of random effect variances, e.g. `Dict(:subject => [0.5, 0.4])` or to specify correlation matrix `Dict(:subject=>[0.5,0.4,I(2,2)],...)`. Technically, this will be passed to the MixedModels.jl `create_re` function, which creates the θ matrices.
@@ -47,7 +47,7 @@ A multiple regression component for one subject.
 All fields can be named. Works best with [`SingleSubjectDesign`](@ref).
 
 # Fields
-- `basis::Any`: an object, if accessed, provides a 'basis-function', e.g. `hanning(40)`, this defines the response at a single event. It will be weighted by the model prediction.
+- `basis::Any`: an object, if accessed, provides a 'basis function', e.g. `hanning(40)::Vector`, this defines the response at a single event. It will be weighted by the model prediction. Future versions will allow for functions, as of v0.3 this is restricted to array-like objects
 - `formula::Any`: StatsModels `formula` object, e.g.  `@formula 0 ~ 1 + cond` (left-hand side must be 0).
 - `β::Vector` Vector of betas/coefficients, must fit the formula.
 - `contrasts::Dict` (optional): Determines which coding scheme to use for which categorical variables. Default is empty which corresponds to dummy coding.
@@ -81,10 +81,10 @@ end
 """
     MultichannelComponent <: AbstractComponent
 
-Wrapper for an `AbstractComponent` to project it to multiple target channels via `projection`.
+Projects a `AbstractComponent` to multiple "channels" via the `projection` vector.
     
 Optionally, `noise` can be added to the source prior to projection.
-A `MultichannelComponent` can be constructed using one of the following options for `projection`:
+By default a `MultichannelComponent` can be constructed using one of the following options for `projection`:
 - `projection::AbstractVector`: Directly pass a custom projection vector.
 - `projection::Pair{<:AbstractHeadmodel,String}`: Generate a projection vector by specifying which headmodel to use and which sources should be active.
 
@@ -178,7 +178,7 @@ maxlength(c::Vector{<:AbstractComponent}) = maximum(length.(c))
 """
     simulate_component(rng, c::AbstractComponent, simulation::Simulation)
 
-By default call `simulate_component` with `(rng, c::Abstractcomponent, design::AbstractDesign)` instead of the whole simulation. This allows users to provide a hook to do something completely different :)
+By default call `simulate_component` with `(rng, c::Abstractcomponent, design::AbstractDesign)` instead of the whole simulation. This function exist solely to provide a "hook" if for a custom component something else than the design is necessary, e.g. a dependency on the onsets, noise or similar.
 """
 simulate_component(rng, c::AbstractComponent, simulation::Simulation) =
     simulate_component(rng, c, simulation.design)
[…]
