### 2024-11-28 5574425
fix sequence design rng
