### 2024-11-28 1966fbe
fix sequence test
