### 2024-11-28 fa6f78b
fix rng docs
