### 2024-12-10 3a37210
Formatted docstrings in design.jl and reorganised the order of the functions in the file.
