### 2024-12-10 bb1da90
Add return value information for generate_events function
