+++ src/design.jl
@@ -217,6 +217,9 @@ Generate a full-factorial events DataFrame based on the experimental conditions 
 # Arguments
 - `design::AbstractDesign`: Experimental design for which the events DataFrame should be created.
 - `rng::AbstractRNG` (optional): Random number generator (RNG) to make the process reproducible. If none is given, `MersenneTwister(1)` will be used.
+
+# Returns
+- `DataFrame`: Each row corresponds to one combination of condition/covariate levels which is often equivalent to one stimulus or trial.
 """
 function generate_events end
 
