### 2024-12-12 9657c93
Deprecated headmodel function and replaced it by an outer constructor for Hartmut
