### 2024-12-12 47a163a
replaced headmodel() with Hartmut() in tests and docs and removed headmodel from exported functions
