### 2024-12-12 b5897d2
replaced headmodel() with Hartmut() in tests and docs and removed headmodel from exported functions
