### 2024-12-12 02e47d6
Apply suggestions from ReviewDog
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
