### 2024-12-12 8ffd9e4
tested nb_muscle.jl
