### 2024-12-12 39351cf
add headmodel again to exported functions
