### 2024-12-13 60b8941
fix #124, explicitly cast the dict
