### 2024-12-16 229dd1e
Added and formatted docstrings in headmodel.jl
