### 2024-12-17 681b5e6
Revise docstrings
