+++ src/headmodel.jl
@@ -103,46 +103,27 @@ Return the leadfield for the (cortical or artefacual) sources of the HArtMuT mod
 julia> h = Hartmut();
 Please cite: HArtMuT: Harmening Nils, Klug Marius, Gramann Klaus and Miklody Daniel - 10.1088/1741-2552/aca8ce
 
-julia> leadfield(h)
-227×2004×3 Array{Float64, 3}:
-[:, :, 1] =
-  0.151429    0.0284341   0.0381777  -0.031533    0.173189    0.00485958   0.0490966   0.0707025   0.00818434  …  -0.029391   -0.480143  -0.140076   -0.405818  -0.0802413  -0.480738    -0.205534   -0.119927     -0.158114
-  0.1732      0.0441432   0.0518934  -0.0101328   0.181351    0.0204942    0.0610806   0.0848026   0.0261072      -0.0259066  -0.48539   -0.0838597  -0.423522  -0.0628599  -0.48615     -0.163832   -0.110515     -0.165316
-  0.249592    0.10857     0.115235    0.10436     0.211371    0.10154      0.125745    0.145194    0.106947        0.0252987  -0.142679   0.373272   -0.20579    0.0663492  -0.243599     0.199307   -0.000593027  -0.0206122
-  0.245206    0.104854    0.110464    0.0929579   0.209626    0.0941604    0.119648    0.140814    0.102867        0.0164751  -0.218093   0.317887   -0.261457   0.0519611  -0.294938     0.155672   -0.0200251    -0.055392
-  0.126496    0.0118467   0.0241024  -0.0535207   0.164124   -0.011041     0.037146    0.0555699  -0.00916642     -0.0319598  -0.465564  -0.190625   -0.380701  -0.0963106  -0.467464    -0.241251   -0.128248     -0.146107
-  ⋮                                                           ⋮                                                ⋱   ⋮                                                         ⋮                                     
- -0.689154   -1.00904    -0.640231   -0.803683   -0.669248   -0.989008    -0.924508   -0.780335   -0.659238       -0.134393   -0.13584   -0.253791   -0.010676  -0.259598    0.00974083  -0.609214   -0.108276     -0.105398
-  0.192729    0.148448    0.131563    0.19185     0.132415    0.164303     0.14867     0.155177    0.157946        0.120672    0.261752   0.189576    0.130332   0.300065    0.144634     0.553918    0.0924756     0.142322
- -1.26181    -1.59936    -1.57901    -1.32513    -1.15289    -1.44583     -1.5914     -1.34047    -0.99371        -0.124783   -0.225718  -0.321099   -0.058764  -0.275479   -0.0440558   -0.679299   -0.140598     -0.133054
-  0.213982    0.145953    0.135357    0.185493    0.155274    0.159222     0.152356    0.160551    0.154056    …   0.123235    0.303627   0.301454    0.13918    0.309234    0.149528     0.621073    0.115515      0.170698
-  0.0731569   0.0794415   0.0814297   0.190411    0.0333774   0.140429     0.105861    0.086699    0.0747807       0.0283169   0.180094  -0.0449803   0.137933  -0.0166636   0.118638     0.0344215   0.0485819     0.107929
+julia> lf = leadfield(h);
 
-[:, :, 2] =Return the leadfield for the (cortical or artefacual) sources of the HArtMuT model.
-  0.591764     0.471349    0.573102    0.478932   0.607152     0.502843    0.472437    0.608999    0.526104     0.540295   …   0.364791    0.712709   0.532255    1.05677    0.335067    1.09307     0.600539   0.418957    0.782385
[…]
