### 2024-12-17 4306147
first draft of leadfield orientation exploration
