+++ docs/make.jl
@@ -53,7 +53,7 @@ makedocs(;
             "Use existing experimental designs & onsets in the simulation" => "./generated/HowTo/predefinedData.md",
         ],
         "Explanations/Explorations" => [
-            "Leadfield Orientations" => "./generated/explanations/orientations_headmodels.jl",
+            "Leadfield Orientations" => "./generated/explanations/orientations_headmodels.md",
         ],
         "API / Docstrings" => "api.md",
     ],
