### 2024-12-17 f527c5a
Update src/noise.jl
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
