### 2025-01-07 #6641 @behinger
happy new year to everyone! @roualdes and @aurorarossi, your checklists are not yet filled in, are you happy with our changes?

Thanks for being patient with us! The push for better docstrings is currently underway, and we think they greatly increase the quality of the software (and we found some minor bugs ;)).
