### 2025-01-09 e2e01e8
revised magnitude function
