+++ joss_paper/paper.bib
@@ -296,7 +296,7 @@ url = {https://doi.org/10.5281/zenodo.8344531}
 
 @software{MATLAB,
 title = {MATLAB},
-author = {The MathWorks Inc.},
+author = {{The MathWorks Inc.}},
 address = {Natick, Massachusetts, United States},
 url = {https://www.mathworks.com}
 }
