### 2025-01-10 a0a2e69
Remove matlab bib entry
