+++ joss_paper/paper.bib
@@ -294,13 +294,6 @@ url = {https://doi.org/10.5281/zenodo.8344531}
 
   % MATLAB
 
-@software{MATLAB,
-title = {MATLAB},
-author = {{The MathWorks Inc.}},
-address = {Natick, Massachusetts, United States},
-url = {https://www.mathworks.com}
-}
-
 @article{vaziri2023eegg,
   title={EEGg Generating Synthetic EEG signals in Matlab environment},
   author={Vaziri, Ava Yektaeian and Makkiabadi, Bahador and Samadzadehaghdam, Nasser},
