+++ joss_paper/paper.md
@@ -42,7 +42,7 @@ One can find a detailed example of how to use `UnfoldSim.jl` to simulate continu
 # Statement of Need
 In our work (e.g. @ehinger2019unfold, @dimigen2021regression), we often analyze data containing (temporally) overlapping events (e.g. stimulus onset and button press, or consecutive eye-fixations), non-linear effects, and complex experimental designs. For a multitude of reasons, we often need to simulate such kind of data: Simulated EEG data is useful to test preprocessing and analysis tools, validate statistical methods, illustrate conceptual issues, test toolbox functionalities, and find limitations of traditional analysis workflows. For instance, such simulation tools allow for testing the assumptions of new analysis algorithms and testing their robustness against any violation of these assumptions.
 
-While other EEG simulation toolboxes exist, they each have limitations: they are dominantly based on the proprietary MATLAB software [@MATLAB], they do not simulate continuous EEG, and they offer little support for designs more complex than two conditions or with non-linear effects. In contrast, `UnfoldSim.jl` is free and open-source and it allows to simulate continuous EEG signals even for complex designs.
+While other EEG simulation toolboxes exist, they each have limitations: they are dominantly based on the proprietary MATLAB software, they do not simulate continuous EEG, and they offer little support for designs more complex than two conditions or with non-linear effects. In contrast, `UnfoldSim.jl` is free and open-source and it allows to simulate continuous EEG signals even for complex designs.
 
 # Functionality
 The package provides four abstract types: `AbstractDesign`, `AbstractComponent`, `AbstractOnset` and `AbstractNoise`. In the following, we present the concrete types that are currently implemented.
