### 2025-01-10 #6641 @jschepers
> Hi @jschepers! It's looking great and we're almost there!
>
> I've read the latest version and have a few formatting changes:
>
>     * Line 29: No reference for MATLAB
>
>     * Line 32, 62: UnfoldSim.jl is not with the same formatting as the abstract
>
>     * Line 55: Not syre if <: is intended to show on the text
>
>     * Figure 1 Legend: "... while the distribution is shown on a linear scale" should perhaps be "... while the uniform distribution is shown on a linear scale"
>
>
> In the comment above, you can see there's additional author tasks, such as archiving on zenodo and posting the DOI here. Please let me know when you've managed to implement the changes and have done the steps above!

Happy New Year everyone!

@mooniean Thank you for the paper suggestions. I revised the paper accordingly:
- I did not include a bibliography entry for MATLAB because I wanted to mention the software as a whole/as programming language and not a specific version or date. I now also removed the information in parantheses to not have an incomplete citation and prevent misunderstandings.
- I unified the formatting of `UnfoldSim.jl`.
- `<:` was supposed to be shown in the text to express a subtype relationship. However, I agree that this might be confusing to the reader therefore I removed it and instead just wrote "a vector of components".
- I agree that this sentence is ambiguous. The word "distribution" was supposed to reference to the display of the lognormal distribution but I now replaced it with "histogram" to make sure that it is clear that it is about the visualization.

I will create a new release/Zenodo version when I finish adapting the docstrings and after potentially adapting the code based on the reviewers' suggestions.
