### 2025-01-10 192f7b3
Update CI.yml
