### 2025-01-10 acec4ff
Update CI.yml
