+++ .github/workflows/CI.yml
@@ -60,8 +60,9 @@ jobs:
             DocMeta.setdocmeta!(UnfoldSim, :DocTestSetup, :(using UnfoldSim); recursive=true)
             doctest(UnfoldSim)'
       - name: Update UnfoldDocs
+        if: github.ref == 'refs/heads/main'
         uses: peter-evans/repository-dispatch@v3
         with:
-          token: ${{ secrets.UNFOLDDOCSPAT }}
-          repository: unfoldtoolbox/UnfoldDocs
-          event-type: individual-docs-updated
+            token: ${{ secrets.UNFOLDDOCSPAT }}
+            repository: unfoldtoolbox/UnfoldDocs
+            event-type: individual-docs-updated
