### 2025-01-13 unfoldtoolbox/UnfoldSim.jl#134 @cbrnr
Do not use `;` in code examples
Just a minor suggestion, if you don't have a good reason to suppress output by appending `;` to a statement, I would suggest to not do that. For example, the slightly longer example in the README should probably not need to use it.

(This issue is part of https://github.com/openjournals/joss-reviews/issues/6641)
