### 2025-01-13 unfoldtoolbox/UnfoldSim.jl#134 @behinger
indeed - I think we used that code in Documenter.jl + Literate.jl , where sometimes not surpressing output results in loooooong dataframes to be plotted in full.

We'll fix it, thanks
