### 2025-01-13 3b3ad59
remove semicolons from readme example
fix #134
