### 2025-01-13 64fd42a
remove semicolons from readme example
fix #134
