### 2025-01-13 unfoldtoolbox/UnfoldSim.jl#134 @cbrnr
Yes, it makes sense to suppress output e.g. in the docs, but maybe try to avoid it at least in code examples?
