### 2025-01-13 #6641 @cbrnr
I've played around with the package a bit, and I think it is really nice! I do have some suggestions to improve certain aspects, so before I'll check the corresponding points in my review list it would be great to see these points addressed:

- [ ] Statement of need: I think it would be nice to sync `README.md` and the documentation homepage with the statement that's currently in the abstract of the manuscript.
- [ ] Installation instructions: `README.md` contains detailed instructions, but I didn't find these in the documentation. Please add them there as well.
- [ ] Automated tests: The package contains several tests, but I think it would be useful to document how to run these tests (maybe right after the installation instructions?).

Aside from that, all other points have been (or are being) addressed, and I'm happy to recommend publication of the manuscript once these last suggestions have been incorporated.
