### 2025-01-13 2b9d4f0
Update magnitude docstring
