### 2025-01-13 50fc4de
adapted magnitude docstring
