### 2025-01-13 f92f8fd
Finished updating headmodel docstrings
