### 2025-01-13 1797a7d
adapted noise.jl docstrings
