+++ src/noise.jl
@@ -1,13 +1,39 @@
 #----------------
-# Types
-#-------------
-
+# Noise types
+#----------------
 
 """
     PinkNoise <: AbstractNoise
-Generate Pink Noise using the SignalAnalysis.jl implementation.
 
-`noiselevel` is used to scale the noise
+A type for generating Pink Noise using the `SignalAnalysis.jl` implementation.
+
+Tip: To manually create noise samples use the [`simulate_noise`](@ref) function.
+
+# Fields
+- `noiselevel = 1` (optional): Factor that is used to scale the noise.
+- `func = SignalAnalysis.PinkGaussian` (optional): Function that is used to create the noise samples.
+    This field is for internal use and should not typically be modified directly by the user.
+    Changes to this field may result in unexpected behavior.
+
+# Examples
+```julia-repl
+julia> noise = PinkNoise(noiselevel = 3)
+PinkNoise
+  noiselevel: Int64 3
+  func: UnionAll
+
+julia> using StableRNGs
+
+julia> simulate_noise(StableRNG(1), noise, 5)
+5-element Vector{Float64}:
+ 2.578878369756878
+ 3.4972108606501786
+ 2.878568584946028
+ 2.2725654770788384
+ 3.5291669151888683
+```
+
+See also [`RedNoise`](@ref), [`WhiteNoise`](@ref), [`ExponentialNoise`](@ref), [`NoNoise`](@ref).
 """
 @with_kw struct PinkNoise <: AbstractNoise
     noiselevel = 1
@@ -17,9 +43,34 @@ end
 
 """
     RedNoise <: AbstractNoise
-Generate Red Noise using the SignalAnalysis.jl implementation.
 
-`noiselevel` is used to scale the noise
+A type for generating Red Noise using the `SignalAnalysis.jl` implementation.
+
+Tip: To manually create noise samples use the [`simulate_noise`](@ref) function.
+
+# Fields
+- `noiselevel = 1` (optional): Factor that is used to scale the noise.
+- `func = SignalAnalysis.RedGaussian` (optional): Function that is used to create the noise samples.
+    This field is for internal use and should not typically be modified directly by the user.
+    Changes to this field may result in unexpected behavior.
+
+# Examples
+```julia-repl
+julia> noise = RedNoise(noiselevel = 2)
+RedNoise
+  noiselevel: Int64 2
+  func: UnionAll
+
+julia> using StableRNGs
+
+julia> simulate_noise(StableRNG(2), noise, 3)
+3-element Vector{Float64}:
+ -0.34153942884005967
+ -0.4651387715669636
+ -0.4951538876376382
+```
+
+See also [`PinkNoise`](@ref), [`WhiteNoise`](@ref), [`ExponentialNoise`](@ref), [`NoNoise`](@ref).
 """
 @with_kw struct RedNoise <: AbstractNoise
     noiselevel = 1
@@ -28,13 +79,34 @@ end
 
 
 """
-    WhiteNoise <: WhiteNoise
-    noiselevel = 1
-    imfilter = 0
-Generate White Noise using `randn` - thus Gaussian noise.
-`noiselevel` is used to scale the noise
+    WhiteNoise <: AbstractNoise
+
+A type for generating White Noise using `randn` - thus Gaussian noise.
 
 Using `imfilter` > 0 it is possible to smooth the noise using Image.imfilter.
+Tip: To manually create noise samples use the [`simulate_noise`](@ref) function.
+
+# Fields
+- `noiselevel = 1` (optional): Factor that is used to scale the noise.
+- `imfilter = 0` (optional):
+
+# Examples
+```julia-repl
[…]
