### 2025-01-15 db9cacb
Revised noise.jl docstrings
