### 2025-01-17 25009e4
Apply suggestions from code review
Co-authored-by: Benedikt Ehinger <benedikt.ehinger@vis.uni-stuttgart.de>
