### 2025-01-17 b1e2c7f
Finished revising onset.jl docstrings
