+++ src/onset.jl
@@ -1,22 +1,57 @@
 #----------------
 # Types
-#---------------
+#----------------
 
 """
-    struct UniformOnset <: AbstractOnset
-Provide a Uniform Distribution of the inter-event-distances.
-`width`  is the width of the uniform distribution (=> the jitter). Since the lower bound is 0, `width` is also the upper bound.
-`offset` is the minimal distance. The maximal distance is `offset + width`.
+    UniformOnset <: AbstractOnset
+
+Provide a Uniform Distribution for the inter-event distances (in samples).
+
+Tip: To manually generate inter-event distance samples use the [`simulate_interonset_distances`](@ref) function.
+
+# Fields
+- `width = 50` (optional): Width of the uniform distribution (=> the "jitter"). Since the lower bound is 0, `width` is also the upper bound.
+- `offset = 0` (optional): The minimal distance between events. The maximal distance is `offset + width`.
+
+# Examples
+```julia-repl
+julia> onset_distribution = UniformOnset(width = 25, offset = 5)
+UniformOnset
+  width: Int64 25
+  offset: Int64 5
+```
+
+See also [`LogNormalOnset`](@ref), [`NoOnset`](@ref).
 """
 @with_kw struct UniformOnset <: AbstractOnset
     width = 50 # how many samples jitter?
     offset = 0 # minimal offset?
 end
 """
-    @with_kw struct LogNormalOnset <: AbstractOnset
-Log-normal inter-event distances using the `Distributions.jl` truncated LogNormal distribution.
+    LogNormalOnset <: AbstractOnset
 
-Be careful with large `μ` and `σ` values, as they are on logscale. σ>8 can quickly give you out-of-memory sized signals!
+Log-normal inter-event distances (in samples) using the `Distributions.jl` truncated LogNormal distribution.
+
+Be careful with large `μ` and `σ` values, as they are on logscale. σ>8 can quickly give you out-of-memory sized signals! \\
+Tip: To manually generate inter-event distance samples use the [`simulate_interonset_distances`](@ref) function.
+
+# Fields
+- `μ`: The mean of the log-transformed variable (the log-normal random variable's logarithm follows a normal distribution).
+- `σ`: The standard deviation of the log-transformed variable.
+- `offset = 0` (optional): The minimal distance between events.
+- `truncate_upper = nothing` (optional): Upper limit (in samples) at which the distribution is truncated.
+
+# Examples
+```julia-repl
+julia> onset_distribution = LogNormalOnset(3, 0.25, 10, 25)
+LogNormalOnset
+  μ: Int64 3
+  σ: Float64 0.25
+  offset: Int64 10
+  truncate_upper: Int64 25
+```
+
+See also [`UniformOnset`](@ref), [`NoOnset`](@ref).
 """
 @with_kw struct LogNormalOnset <: AbstractOnset
     μ::Any  # mean
@@ -26,16 +61,67 @@ Be careful with large `μ` and `σ` values, as they are on logscale. σ>8 can qu
 end
 
 """
-    struct NoOnset <: AbstractOnset end
-In the case that the user directly wants no overlap to be simulated (=> epoched data).
+    NoOnset <: AbstractOnset
+
+For cases where the user wants to simulate epoched data without any overlap between consecutive events.
+
+# Examples
+```julia-repl
+julia> onset_distribution = NoOnset()
+NoOnset()
+```
+
+See also [`UniformOnset`](@ref), [`LogNormalOnset`](@ref).
 """
 struct NoOnset <: AbstractOnset end
 
+
+#-----------------------------
+# Onset simulation functions
+#-----------------------------
+
 """
-    simulate_interonset_distances(rng, onset::UniformOnset, design::AbstractDesign)
[…]
