### 2025-01-22 #6641 @roualdes
Great work, y'all.  Looks good.  In the paper, I think the word kind needs an s on it, but that's the only thing I have to offer. I recommend publication. Cheers
