### 2025-01-24 b1655b5
Revised docstrings for predefined simulations
