+++ src/predefinedSimulations.jl
@@ -1,42 +1,85 @@
-# here we can define some predefined Simulations. Convenient if you just want to have a quick simulation :)
+# Here we define some predefined simulations. Convenient if you just want to have a quick simulation :)
 
-
-
+# If no RNG is given, create one.
 predef_2x2(; kwargs...) = predef_2x2(MersenneTwister(1); kwargs...) # without rng always call same one
 predef_eeg(; kwargs...) = predef_eeg(MersenneTwister(1); kwargs...) # without rng always call same one
 predef_eeg(nsubjects::Int; kwargs...) = predef_eeg(MersenneTwister(1), nsubjects; kwargs...) # without rng always call same one
 
+
 """
-    predef_eeg(; kwargs...)
-    predef_eeg(rng; kwargs...)
-    predef_eeg(rng, n_subjects; kwargs...)
+    predef_eeg(; <keyword arguments>)
+    predef_eeg(rng; <keyword arguments>)
+    predef_eeg(rng, n_subjects; <keyword arguments>)
 
-Generate a P1/N1/P3 complex.
-In case `n_subjects` is defined - `MixedModelComponents` are generated, else `LinearModelComponents`.
+Simulate data for a P1/N1/P3 component complex.
 
-The most used `kwargs` is: `return_epoched=true` which returns already epoched data. If you want epoched data without overlap, specify `onset = NoOnset()` and `return_epoched = true`
+Note that this function is mainly used for demonstration and internal testing purposes or whenever a quick simulation is needed. \n
+In case `n_subjects` is defined - `MixedModelComponents` are generated (multi-subject simulation), else `LinearModelComponents` (single-subject simulation). \n
+The most used keyword argument is: `return_epoched = true` which returns already epoched data. If you want epoched data without overlap, specify `onset = NoOnset()` and `return_epoched = true`.
 
+# Keyword arguments
 
-## Default parameters:
+## Design
+- `n_repeats = 100`: Number of times the experimental design is repeated. Only used in the single-subject case.
+- `event_order_function = shuffle`: Random trial order.
+- `conditions = Dict(:condition => ["car", "face"], :continuous => range(-5, 5, length = 10))`: Conditions and covariates used in the design.
 
-#### Design
-- `n_repeats = 100`,
-- `event_order_function = shuffle`, # random trial order
-- `conditions = Dict(...)`,
+## Component / Signal
+- `sfreq = 100`: Sampling frequency.
+- `p1 = (p100(; sfreq = sfreq), @formula(0 ~ 1), [5], Dict())`: P1 with amplitude 5; no effects.
+- `n1 = (n170(; sfreq = sfreq), @formula(0 ~ 1 + condition), [5,-3], Dict())`: N1 with amplitude 5, dummy-coded condition effect (levels "car", "face") of -3.
+- `p3 = (p300(; sfreq = sfreq), @formula(0 ~ 1 + continuous), [5,1], Dict())`: P3 with amplitude 5, continuous effect range [-5,5] with slope 1.
 
-#### Component / Signal
-- `sfreq = 100`,
-- `p1 = (p100(; sfreq = sfreq), @formula(0 ~ 1), [5], Dict())`, # P1 amp 5, no effects
-- `n1 = (n170(; sfreq = sfreq), @formula(0 ~ 1 + condition), [5,-3], Dict())`, # N1 amp 5, dummy-coded condition effect (levels "car", "face") of -3
-- `p3 = (p300(; sfreq = sfreq), @formula(0 ~ 1 + continuous), [5,1], Dict())`, # P3 amp 5, continuous effect range [-5,5] with slope 1
-
-#### Onset
+## Onset
[…]
