### 2025-01-27 266c8d8
Apply suggestions from code review
Co-authored-by: Benedikt Ehinger <benedikt.ehinger@vis.uni-stuttgart.de>
