### 2025-01-28 ee6376c
Update src/predefinedSimulations.jl docstrings
Co-authored-by: Benedikt Ehinger <benedikt.ehinger@vis.uni-stuttgart.de>
