### 2025-01-28 0628bee
Removed convert function since it was obsolete in UnfoldSim
