### 2025-01-28 80b30eb
update docstrings simulate.jl
