### 2025-01-28 a65823f
Added docstring for Simulation type
