### 2025-01-30 b1f456d
update docstring examples in simulation.jl
