### 2025-01-31 ac31308
Added docstrings for helper.jl
