+++ src/helper.jl
@@ -1,9 +1,57 @@
 """
-Pads array with specified value, length
-pad_array(arr, len, val)
+    pad_array(arr::Vector, len::Int, val)    
+    pad_array(arr::Vector, len::Tuple, val)
+
+Pads the input array `arr` with a specified value `val` either before or after the existing elements, based on the sign of `len`.
+
+# Arguments
+- `arr::Vector`: The input array to be padded.
+- `len::Union{Int, Tuple}`: The number of times that `val` should be added.
+    If `len` is negative, the values are added before the existing array. Otherwise, they are added after the existing array.
+    If `len` is a tuple, `pad_array` is called twice which enables padding the array before and after in one function call.
+- `val`: The value to be used for padding.
+
+# Returns
+- `Vector`: Padded vector with the new length `length(arr) + sum(abs.(len))`.
+
+# Examples
+```julia-repl
+# Create an array that will be padded
+julia> my_array = rand(5)
+5-element Vector{Float64}:
+ 0.0420017254437951
+ 0.19179144973603235
+ 0.5388760239550549
+ 0.6973699906283798
+ 0.9966598131018376
+
+ # Pad the array with zeros before the original array
+julia> pad_array(my_array, -2, 0)
+7-element Vector{Float64}:
+ 0.0
+ 0.0
+ 0.0420017254437951
+ 0.19179144973603235
+ 0.5388760239550549
+ 0.6973699906283798
+ 0.9966598131018376
+
+# Pad the array with the value 5 before and after the original array
+julia> pad_array(my_array, (-2, 1), 5)
+8-element Vector{Float64}:
+ 5.0
+ 5.0
+ 0.0420017254437951
+ 0.19179144973603235
+ 0.5388760239550549
+ 0.6973699906283798
+ 0.9966598131018376
+ 5.0
+```
 """
 pad_array(arr::Vector, len::Tuple, val) =
     pad_array(pad_array(arr, len[1], val), len[2], val)
+
 function pad_array(arr::Vector, len::Int, val)
     pad = fill(val, abs(len))
     arr = len > 0 ? vcat(arr, pad) : vcat(pad, arr)
@@ -12,14 +60,37 @@ end
 
 
 """
-	closest_src(coords_list::AbstractVector{<:AbstractVector}, pos)
-	closest_src(coords::Vector{<:Real}, pos) 
-	
-Takes an array of 'm' target coordinate vector (size 3) (or vector of vectors) and a matrix (n-by-3) of all available positions, and returns an array of size 'm' containing the indices of the respective items in 'pos' that are nearest to each of the target coordinates.
-"""
-closest_src(coords_list::AbstractVector{<:AbstractVector}, pos) =
-    closest_src.(coords_list, Ref(pos))
+	closest_src(coords::Vector{<:Real}, pos)
 
+Return the index of the position that is closest to the target coordinates (using the Euclidean distance).
+
+# Arguments
+- `coords::Vector{<:Real}`: Target coordinate vector (typically with length 3).
+- `pos`: Matrix that contains all possible positions.
+    Its dimensions are the number of positions `n` times number of entries in the position coordinate vectors (usually 3) i.e. `n x 3`.
+
+# Returns
+- `Int`: Row index of the position in `pos` that is closest to `coords`.
+
+# Examples
+```julia-repl
+julia> positions = [0 0 0; 2 2 2; 3 3 3; 4 4 4]
+4×3 Matrix{Int64}:
+ 0  0  0
+ 2  2  2
+ 3  3  3
+ 4  4  4
+
+julia> target_coordinates = [0.5, 0, 0.5]
+3-element Vector{Float64}:
+ 0.5
+ 0.0
+ 0.5
+
[…]
