+++ src/simulation.jl
@@ -30,15 +30,15 @@ end
     return_epoched = false,
     )
 
-Return continuous or epoched signal, given `Design`, [Array of] `Component`, `Onset` and 
-optional [`Noise`] and [`RNG`]. Main simulation function.
+Simulate continuous or epoched signal, given `design`, [Array of] `component`, `onset` and 
+optional `noise` and `rng`. Main simulation function.
 
 # Arguments
-- `design::AbstractDesign`: Desired experimental design.
-- `signal`: `Component` for the desired signal.
-- `onset::AbstractOnset`: Desired onset.
-- `noise::AbstractNoise = NoNoise()` (optional): Desired noise.
 - `rng::AbstractRNG` (optional): Random number generator, important to ensure reproducibility.
+- `design::AbstractDesign`: Desired experimental design.
+- `signal`: `Component`(s) for the desired signal.
+- `onset::AbstractOnset`: Desired inter-onset distance distribution.
+- `noise::AbstractNoise = NoNoise()` (optional): Desired noise.
 
 # Keyword arguments
 - `return_epoched = false`: Skip the Onset-calculation and conversion to continuous data 
@@ -69,21 +69,34 @@ julia> onset = UniformOnset(; width = 20, offset = 4);
 
 julia> noise = PinkNoise(; noiselevel = 0.2);
 
-julia> data, events = simulate(MersenneTwister(1), design, signal, onset, noise)
-([-0.045646938524459196, 0.15784406738265955, 0.012640319497460443, 0.026669512219327673, 0.15329144053662508, 0.06412654786607011, -0.16766777448918685, -0.08012027590515228, 0.0020515088981202137, -0.24874482217391175  …  0.24621397283439814, 0.1710771262918883, -0.01527736524528042, 0.7639978745937471, 1.5600315092771557, 1.624219837479329, 1.2889713838347956, 0.26819223928179, 0.16535758767503866, 0.21291936972924855], 20×2 DataFrame
+# Variant 1: Use a custom RNG.
+julia> data, events = simulate(MersenneTwister(2), design, signal, onset, noise);
+
+julia> data
+293-element Vector{Float64}:
+ -0.013583193323430123
+  0.09159433856866195
+  ⋮
+ -0.25190584567097907
+ -0.20179992275876316
+
+julia> events
+20×2 DataFrame
  Row │ cond_A   latency 
      │ String   Int64   
 ─────┼──────────────────
-   1 │ level_A       18
-   2 │ level_B       39
-   3 │ level_A       45
-   4 │ level_B       49
+   1 │ level_A        9
+   2 │ level_B       20
+   3 │ level_A       27
+   4 │ level_B       37
   ⋮  │    ⋮        ⋮
-  18 │ level_B      258
-  19 │ level_A      281
-  20 │ level_B      303
+  18 │ level_B      257
+  19 │ level_A      271
+  20 │ level_B      284
+         13 rows omitted
 
-julia> data1, events1 = simulate(design, signal, onset, noise)
+# Variant 2: Without specifying an RNG, MersenneTwister(1) will be used for the simulation.
+julia> data1, events1 = simulate(design, signal, onset, noise);
 ┌ Warning: No random generator defined, used the default (`Random.MersenneTwister(1)`) with a fixed seed. This will always return the same results and the user is strongly encouraged to provide their own random generator!
 ```
 
@@ -164,7 +177,7 @@ end
 """
     create_continuous_signal(rng, responses, simulation)
 
-Simulate onset latencies and add together a continuous signal, based on the given responses and simulation parameters.
[…]
