### 2025-01-31 8d84817
Added information to the simulate docstring
