### 2025-02-01 9cc957f
Apply suggestions from code review
Co-authored-by: Judith Schepers <judith.schepers@vis.uni-stuttgart.de>
