### 2025-02-03 f850674
Renamed potentially confusing variables
