+++ src/simulation.jl
@@ -7,16 +7,22 @@ Simulation(
 ) = Simulation(design, [component], onset, noisetype)
 
 
-function simulate(design::AbstractDesign, signal, onset::AbstractOnset, args...; kwargs...)
+function simulate(
+    design::AbstractDesign,
+    components,
+    onset::AbstractOnset,
+    args...;
+    kwargs...,
+)
     @warn "No random generator defined, used the default (`Random.MersenneTwister(1)`) with a fixed seed. This will always return the same results and the user is strongly encouraged to provide their own random generator!"
-    simulate(MersenneTwister(1), design, signal, onset, args...; kwargs...)
+    simulate(MersenneTwister(1), design, components, onset, args...; kwargs...)
 end
 
 """
     simulate(
     rng::AbstractRNG,
     design::AbstractDesign,
-    signal,
+    components,
     onset::AbstractOnset,
     noise::AbstractNoise = NoNoise();
     return_epoched = false,
@@ -24,7 +30,7 @@ end
 
 	simulate(
     design::AbstractDesign,
-    signal,
+    components,
     onset::AbstractOnset,
     noise::AbstractNoise = NoNoise();
     return_epoched = false,
@@ -36,7 +42,7 @@ optional `noise` and `rng`. Main simulation function.
 # Arguments
 - `rng::AbstractRNG` (optional): Random number generator, important to ensure reproducibility.
 - `design::AbstractDesign`: Desired experimental design.
-- `signal`: `Component`(s) for the desired signal.
+- `components`: `Component`(s) for the desired signal.
 - `onset::AbstractOnset`: Desired inter-onset distance distribution.
 - `noise::AbstractNoise = NoNoise()` (optional): Desired noise.
 
@@ -59,7 +65,7 @@ julia> design =
     SingleSubjectDesign(; conditions = Dict(:cond_A => ["level_A", "level_B"])) |>
     x -> RepeatDesign(x, 10);
 
-julia> signal = LinearModelComponent(; 
+julia> component = LinearModelComponent(; 
     basis = [0, 0, 0, 0.5, 1, 1, 0.5, 0, 0],
     formula = @formula(0 ~ 1 + cond_A),
     β = [1, 0.5],
@@ -70,7 +76,7 @@ julia> onset = UniformOnset(; width = 20, offset = 4);
 julia> noise = PinkNoise(; noiselevel = 0.2);
 
 # Variant 1: Use a custom RNG.
-julia> data, events = simulate(MersenneTwister(2), design, signal, onset, noise);
+julia> data, events = simulate(MersenneTwister(2), design, component, onset, noise);
 
 julia> data
 293-element Vector{Float64}:
@@ -96,7 +102,7 @@ julia> events
          13 rows omitted
 
 # Variant 2: Without specifying an RNG, MersenneTwister(1) will be used for the simulation.
-julia> data1, events1 = simulate(design, signal, onset, noise);
+julia> data1, events1 = simulate(design, component, onset, noise);
 ┌ Warning: No random generator defined, used the default (`Random.MersenneTwister(1)`) with a fixed seed. This will always return the same results and the user is strongly encouraged to provide their own random generator!
 ```
 
@@ -115,11 +121,11 @@ Additional remarks on the overlap of adjacent signals when `return_epoched = tru
 simulate(
     rng::AbstractRNG,
     design::AbstractDesign,
-    signal,
+    components,
     onset::AbstractOnset,
     noise::AbstractNoise = NoNoise();
     kwargs...,
-) = simulate(rng, Simulation(design, signal, onset, noise); kwargs...)
[…]
