+++ docs/run_liveserver.jl
@@ -5,6 +5,7 @@ push!(dr, "./build")
 servedocs(
     skip_dirs = dr,
     literate_dir = joinpath("literate"),
+    include_dirs = [joinpath("..", "src")],
     foldername = ".",
     host = "0.0.0.0",
 )
