### 2025-02-04 23d2bc8
Implemented reviewer suggestion
