+++ docs/src/index.md
@@ -2,30 +2,105 @@
 CurrentModule = UnfoldSim
 ```
 
-# UnfoldSim.jl
+# UnfoldSim.jl Documentation
 
-Documentation for [UnfoldSim.jl](https://github.com/unfoldtoolbox/UnfoldSim.jl): a Julia package for simulating multivariate timeseries data with a special focus on EEG data.
+Welcome to [UnfoldSim.jl](https://github.com/unfoldtoolbox/UnfoldSim.jl): a Julia package for simulating multivariate timeseries data with a focus on EEG, especially event-related potentials (ERPs). 
+The user provides four ingredients: 1) an experimental design, with both categorical and continuous variables, 2) event basis functions specified via linear or hierarchical models, 3) an inter-event onset distribution, and 4) a noise specification.
 
-## Start simulating time series data
-We offer some predefined (EEG) signals, check them out!
+```@raw html
+<div style="width:60%; margin: auto;">
 
-For instance a P1/N170/P300 complex (containing three typical ERP components).
-```@example main
-using UnfoldSim
-using CairoMakie # plotting
-
-data, evts = UnfoldSim.predef_eeg(; n_repeats = 1, noiselevel = 0.8)
-
-lines(data; color = "black")
-vlines!(evts.latency; color = ["orange", "teal"][1 .+ (evts.condition.=="car")])
-
-current_figure()
+<img src="assets/UnfoldSim_features_animation.gif"/>
+</div>
 ```
 
-## Or simulate epoched data directly
-```@example main
+## Key features
+- **Modularity:** Choose or implement different designs, components, onset distributions or noise types
+- **Multi-subject** & complex experimental designs
+- **Multi-channel** via EEG-forward models
+- **Continuous or epoched** data with potentially **overlapping** signals
+- Potential support for **other modalities**, e.g. single-voxel fMRI or pupil dilation
 
-data, evts = UnfoldSim.predef_eeg(; n_repeats = 20, noiselevel = 0.8, return_epoched = true)
-heatmap(data[:, sortperm(evts, [:condition, :continuous])])
+## Installation
+```julia-repl
+julia> using Pkg; Pkg.add("UnfoldSim")
+```
 
+## Usage example
+#### Start simulating time series data
+We offer some predefined (EEG) signals, check them out!
+For instance, a P1/N170/P300 complex (containing three typical ERP components).
+
+```@example quickstart
+using UnfoldSim
+using CairoMakie # plotting #hide
+data, events = UnfoldSim.predef_eeg(; n_repeats = 1, noiselevel = 0.8)
+
+f = Figure(size = (900, 400)) #hide
+ax = Axis( #hide
+    f[1, 1], #hide
+    title="Simulated EEG data", #hide
+    titlesize=18, #hide
+    xlabel="Time [samples]", #hide
+    ylabel="Amplitude [µV]", #hide
+    xlabelsize=16, #hide
+    ylabelsize=16, #hide
+    xgridvisible=false, #hide
+    ygridvisible=false, #hide
+) #hide
+
+colors = ifelse.(events.condition .== "face", :orange , :teal) #hide
+
+lines!(data; color="black") #hide
+vlines!(events.latency; color=colors, label=events.condition) #hide
+
+xlims!(current_axis(), [0, 400]) #hide
+
+current_figure() #hide
+```
+
+#### Or simulate epoched data directly
+```@example quickstart
+
+data, events = UnfoldSim.predef_eeg(; n_repeats = 20, noiselevel = 0.8, return_epoched = true)
+
+f2 = Figure(size = (900, 400)) #hide
+ax = Axis( #hide
+    f2[1, 1], #hide
[…]
