### 2025-02-07 8e48c3c
Start revising README: Added statement of need
