### 2025-02-07 89a72bd
Adapt installation instructions in the README
