### 2025-02-07 6763b83
Added installation instructions based on the ones from Unfold, thanks @behiger
