### 2025-02-07 e60bf1c
Added installation instructions based on the ones from Unfold, thanks @behiger
