### 2025-02-07 241cb0e
Remove duplicate section
