### 2025-02-11 8f23c26
Test adding animation to README
