### 2025-02-11 dda5df9
Adapted animation gif link and deleted previous image
