+++ README.md
@@ -15,9 +15,7 @@ UnfoldSim.jl provides multi-channel support via EEG-forward models. Moreover, it
 
 Many tutorials, guides, how-tos and references are available in the [documentation](https://unfoldtoolbox.github.io/UnfoldSim.jl/)!
 
-![readme_figure](https://github.com/unfoldtoolbox/UnfoldSim.jl/assets/22366977/b69d186c-fd3d-4449-9f2e-03d7e01b8cb3)
-
-![unfoldsim_animation](docs/assets/UnfoldSim_features_animation.gif)
+![unfoldsim_animation](https://github.com/unfoldtoolbox/UnfoldSim.jl/blob/assets/docs/src/assets/UnfoldSim_features_animation.gif)
 
 ## Installation
 
