### 2025-02-11 b0a8666
Minor changes to README and docs landing page
