### 2025-02-11 da81976
Minor changes to README and docs landing page
