+++ docs/src/developer_docs.md
@@ -1,0 +1,242 @@
+# Developer documentation
+
+Welcome to the developer documentation! We are excited that you are interested in contributing to our package.
+
+Feel free to submit your work in a state you are comfortable with—we genuinely appreciate every contribution! If you are interested in following best practices and learning along the way, keep reading. But don't worry, we welcome your input just as it is 🙂.
+
+
+!!! note "Contribution guide"
+    If you haven't already, please read the [Contribution guide](https://github.com/unfoldtoolbox/UnfoldSim.jl/blob/main/CONTRIBUTING.md) first.
+
+Please note that the following documentation is adapted from the [BestieTemplate.jl developer documentation](https://juliabesties.github.io/BestieTemplate.jl/stable/91-developer/) but has been customized to fit our needs.
+
+## Development/GitHub Workflow
+
+### Before you start coding
+1. Check whether there exists a GitHub issue about the topic (e.g. bug, new feature etc). If not create one with a short description of the problem or feature idea.
+2. Discuss your approach with the package maintainers either in the issue or via another channel.
+
+### First time clone & development version
+
+If this is the first time you work with this repository, follow the instructions below to clone the repository and create a `dev` version. 
+
+!!! tip "`dev` version of a Julia package"
+    Having a `dev` (development) version of a package allows you to import a local version of the package with your changes instead of the registered package version (which is static).
+
+#### a) If you have writing access for the GitHub repository
+- Option 1: Clone this repository using `git clone`.
+- Option 2 (recommended): Use the Julia `dev` command to create a development version of the package:
+    1. Start a Julia session and run `cd("/path/to/your/project")` to navigate to your project folder.
+    2. Press `]` to enter `pkg` mode.
+    3. Run `dev --local UnfoldSim` to clone the package to `./dev/UnfoldSim` and automatically add it to your Julia project environment.
+
+!!! important
+    If you have writing rights, whenever **upstream** is mentioned, use **origin** instead.
+
+#### b) If you don't have writing access for the GitHub repository
+1. Fork this repository.
+2. Clone your repository (this will create a `git remote` called `origin`).
+3. Add this repository as a remote:
+
+   ```bash
+   git remote add upstream https://github.com/unfoldtoolbox/UnfoldSim.jl
+   ```
+
+This will ensure that you have two remotes in your git: `origin` and `upstream`.
+You will create branches and push to `origin`, and you will fetch and update your local `main` branch from `upstream`.
+
+!!! tip "`dev` version without writing rights"
+    You can also use the `dev` command on your fork. Run `]dev --local url/of/your/fork` to clone the package to `./dev/UnfoldSim` and automatically add it to your Julia project environment.
+
+### Revise.jl
+Further, we recommend to use [`Revise.jl`](https://github.com/timholy/Revise.jl): a Julia package which allows you to track source code changes in a running Julia session without need to restart it and reload the package.
+
+We recommend to install it in the global environment:
+```julia-repl
+julia> # Press ]
+pkg> activate
+pkg> add Revise
+```
+
+### Working on a new issue
+
+We try to keep a linear history in this repo, so it is important to keep your branches up-to-date.
+
+1. Fetch from the remote and fast-forward your local main
+
+   ```bash
+   git fetch upstream
+   git switch main
+   git merge --ff-only upstream/main
+   ```
+
+2. Branch from `main` to address the issue (see below for naming)
+
+   ```bash
+   git switch -c 42-add-answer-universe
+   ```
+
+3. Push the new local branch to your personal remote repository
+
+   ```bash
[…]
