### 2025-02-13 ffc7094
add: Sequential Sampling Model, Evidence Accumulation Simulation
