### 2025-02-18 bf2e7c5
Minor revisions to the developer documentation page
