+++ docs/literate/HowTo/simulateDriftOverlap.jl
@@ -11,50 +11,58 @@
 # ```
 using UnfoldSim
 using Unfold
-using Random
+using StableRNGs
 using CairoMakie, UnfoldMakie
+using SequentialSamplingModels
 
 fs = 500
 Δt = 1/fs; # time step
 tEnd = 1.0 # trial Duration
 time_vec = 0:Δt:tEnd; # time base - let's make it a typical stimulus duration
 
-design_single = SingleSubjectDesign()
-design_seq = SequenceDesign(design_single, "SCR_")
-design_rep = RepeatDesign(design_seq, 500)
 p3 = LinearModelComponent(;
-    basis = UnfoldSim.hanning(Int(0.5 * fs)),
+    basis = UnfoldSim.hanning(Int(0.7 * fs)),
     formula = @formula(0 ~ 1 + condition),
-    β = [1.0, 0],
+    β = [0.8, 0],
 )
 resp = LinearModelComponent(;
-    basis = UnfoldSim.hanning(Int(0.5 * fs)),
+    basis = UnfoldSim.hanning(Int(0.7 * fs)),
     formula = @formula(0 ~ 1 + condition),
     β = [0.5, 0],
 )
-v = 0.6 # Drift Rate
-a = 2.0 # The amount of information that is considered for a decision.
-t = 0.10 # The duration for a non-decisional processes (encoding and response execution).
-z = 0.50 # An indicator of an an initial bias towards a decision.
-ddm_parameter = Dict(:ν=>v, :α=>a, :z=>z, :τ=>t)
-drift = Drift_Component(
-    simulate_component,
-    time_vec,
-    Δt,
-    DDM,
-    ddm_parameter)
-components = Dict('S' => [p3], 'C' => [drift], 'R' => [resp])
 
 # ```@raw html
 # </details >
 # ```
+# ## Design
+# Let's generate a single design with two different drift_rates as condition, so we can use them later for the Simulation
+design_single = SingleSubjectDesign(
+    conditions = Dict(
+        :drift_rate => [3, 5],
+        :condition => [1])
+)
+design_seq = SequenceDesign(design_single, "SCR_")
+design_rep = RepeatDesign(design_seq, 500)
 
+# ## Create the Drift Component using the special KellyModel for the evidence accumulation
+# As parameter for the models drift_rate we use the drift_rate specified in the design by directly naming the condition as string.
+v = "drift_rate" # Drift Rate from the design
+# Now we retrieve the models default parameters, but we change the drift_rate and
+kelly_model = KellyModel(drift_rate=v, motor_onset=0.4, event_onset=0.2)
+kelly_model_parameter = UnfoldSim.create_kelly_parameters_dict(kelly_model)
+drift = Drift_Component(
+    simulate_component,
+    time_vec,
+    Δt,
+    KellyModel,
+    kelly_model_parameter)
+components = Dict('S' => [p3], 'C' => [drift], 'R' => [resp])
 # ## Create the Onset Component to simulate the overlap
 # To simulate an overlap between the drift_component in 'C' and the response component in 'R'. We have to specify the UniformOnset of the 'C' Component therefore with an negative offset to produce the overlap.
 seq_onset = SequenceOnset(
-    Dict('S'=>UniformOnset(width=0,offset=250),
-         'C'=>(DriftOnset(), UniformOnset(width=0, offset=-150)),
-         'R'=>UniformOnset(width=0,offset=300)))
[…]
