### 2025-02-19 #6641 @jschepers
Hello everyone,

I finally found the time to address the remaining reviewer suggestions and we completed revising all UnfoldSim.jl docstrings (unfoldtoolbox/UnfoldSim.jl#121).

@roualdes Thanks for reading the paper again. I added the missing 's'.
@cbrnr Thanks again for your suggestions regarding the README and the package documentation. We implemented them in unfoldtoolbox/UnfoldSim.jl#138.

Next, I will release a new package version and complete the additional author tasks after review.

Best,
Judith
