### 2025-02-20 c524e9b
Update CITATION.cff
