+++ test/design.jl
@@ -12,11 +12,6 @@
         ) == 1
         des = SingleSubjectDesign(;
             conditions = Dict(:A => nlevels(5), :B => nlevels(2)),
-<<<<<<< HEAD
-            event_order_function = x -> sort(x, order(:B, rev = true)),
-        )
-        @test generate_events(des).B[1] == "S2"
-=======
             event_order_function = (rng, x) -> sort(x, order(:B, rev = true)),
         )
         @test generate_events(des).B[1] == "S2"
@@ -31,7 +26,6 @@
         # different sortig seed results in different sorts
         @test generate_events(MersenneTwister(3), des) !=
               generate_events(MersenneTwister(4), des)
->>>>>>> main
     end
 
     @testset "MultiSubjectDesign" begin
@@ -52,11 +46,7 @@
             n_subjects = 10,
             n_items = 100,
             both_within = Dict(:A => nlevels(5), :B => nlevels(2)),
-<<<<<<< HEAD
-            event_order_function = x -> sort(x, order(:item, rev = true)),
-=======
             event_order_function = (rng, x) -> sort(x, order(:item, rev = true)),
->>>>>>> main
         )
         @test generate_events(des).subject[1] == "S01"
 
@@ -65,11 +55,6 @@
             n_subjects = 10,
             n_items = 100,
             both_within = Dict(:A => nlevels(5), :B => nlevels(2)),
-<<<<<<< HEAD
-            event_order_function = x -> sort(x, order(:B, rev = true)),
-        )
-        @test generate_events(des).B[1] == "S2"
-=======
             event_order_function = (rng, x) -> sort(x, order(:B, rev = true)),
         )
         @test generate_events(des).B[1] == "S2"
@@ -88,7 +73,6 @@
         # different sortig seed results in different sorts
         @test generate_events(MersenneTwister(3), des) !=
               generate_events(MersenneTwister(4), des)
->>>>>>> main
 
         # check that this throws an error because of `dv` as condition name
         des = MultiSubjectDesign(;
@@ -204,8 +188,4 @@
         @test length(unique(i1.cond)) == 1
     end
 
-<<<<<<< HEAD
 end
-=======
-end
->>>>>>> main
