+++ src/design.jl
@@ -1,23 +1,7 @@
-<<<<<<< HEAD
-""" 
-    MultiSubjectDesign
-
-- `n_subjects`::Int -> number of subjects
-- `n_items`::Int -> number of items (sometimes ≈trials)
-- `subjects_between` = Dict{Symbol,Vector} -> effects between subjects, e.g. young vs old 
-- `items_between` = Dict{Symbol,Vector} -> effects between items, e.g. natural vs artificial images, (but shown to all subjects if not specified also in `subjects_between`)
-- `both_within` = Dict{Symbol,Vector}	-> effects completly crossed
-- `event_order_function` = `x->x`; # can be used to sort, or e.g. `x->shuffle(MersenneTwister(42),x)` - be sure to fix/update the rng accordingly!!
-
-tipp: check the resulting dataframe using `generate_events(design)`
-
-
-=======
 # Design types
 
 """
     SingleSubjectDesign <: AbstractDesign
->>>>>>> main
 
 A type for specifying the experimental design for a single subject (based on the given conditions).
 
@@ -125,38 +109,18 @@ See also [`SingleSubjectDesign`](@ref), [`RepeatDesign`](@ref)
     subjects_between::Dict{Symbol,Vector} = Dict()
     items_between::Dict{Symbol,Vector} = Dict()
     both_within::Dict{Symbol,Vector} = Dict()
-<<<<<<< HEAD
-    event_order_function = x -> x # can be used to sort, or x->shuffle(rng,x)
-=======
     event_order_function = (rng, x) -> x
->>>>>>> main
 end
 
 """
-<<<<<<< HEAD
-- conditions = Dict{Symbol,Vector} of conditions, e.g. `Dict(:A=>["a_small","a_big"],:B=>["b_tiny","b_large"])`
-- `event_order_function` = x->x; # can be used to sort, or x->shuffle(MersenneTwister(42),x) - be sure to fix/update the rng accordingly!!
-
-Number of trials / rows in `generate_events(design)` depend on the full factorial of your `conditions`.
-=======
     RepeatDesign{T} <: AbstractDesign
 
 Repeat a design (and the corresponding events DataFrame) multiple times to mimick repeatedly recorded trials.
->>>>>>> main
 
 Tip: Check the resulting dataframe using the [`generate_events`](@ref) function.
 Please note that when using an `event_order_function`(e.g. `shuffle`) in a `RepeatDesign`, the corresponding RNG is shared across repetitions and not deep-copied for each repetition.
 As a result, the order of events will differ for each repetition.
 
-<<<<<<< HEAD
-If conditions are omitted (or set to `nothing`), a single trial is simulated with a column `:dummy` and content `:dummy` - this is for convenience.
-
-tipp: check the resulting dataframe using `generate_events(design)`
-"""
-@with_kw struct SingleSubjectDesign <: AbstractDesign
-    conditions::Dict{Symbol,Vector} = Dict()
-    event_order_function = x -> x
-=======
 # Fields
 - `design::T`: The experimental design that should be repeated.
 - `repeat::Int = 1`: The number of repetitions.
@@ -209,31 +173,11 @@ See also [`SingleSubjectDesign`](@ref), [`MultiSubjectDesign`](@ref)
 @with_kw struct RepeatDesign{T} <: AbstractDesign
     design::T
     repeat::Int = 1
->>>>>>> main
 end
 
 #----
 # Design helper functions
 
-<<<<<<< HEAD
-""" Returns dimension of experiment design"""
-size(design::MultiSubjectDesign) = (design.n_items, design.n_subjects)
-size(design::SingleSubjectDesign) = (*(length.(values(design.conditions))...),)
-
-"""
-Generates full-factorial DataFrame of design.conditions
-
[…]
