### 2025-02-24 #6641 @jschepers
Dear @mooniean,

we released a new package version with all the changes from the review:

Version: v0.4.0
Zenodo: 10.5281/zenodo.14894630

Is there anything else that still needs to be done?

Best,
Judith
