### 2025-02-25 a0f9253
Update src/component.jl
Wording DocString

Co-authored-by: Benedikt Ehinger <benedikt.ehinger@vis.uni-stuttgart.de>
