### 2025-02-25 ff43378
Update src/component.jl
Co-authored-by: Benedikt Ehinger <benedikt.ehinger@vis.uni-stuttgart.de>
