+++ docs/literate/HowTo/simulateDriftOverlap.jl
@@ -16,7 +16,7 @@ using CairoMakie, UnfoldMakie
 using SequentialSamplingModels
 
 fs = 500
-Δt = 1/fs; # time step
+Δt = 1 / fs; # time step
 tEnd = 1.0 # trial Duration
 time_vec = 0:Δt:tEnd; # time base - let's make it a typical stimulus duration
 
@@ -36,11 +36,8 @@ resp = LinearModelComponent(;
 # ```
 # ## Design
 # Let's generate a single design with two different drift_rates as condition, so we can use them later for the Simulation
-design_single = SingleSubjectDesign(
-    conditions = Dict(
-        :drift_rate => [3, 5],
-        :condition => [1])
-)
+design_single =
+    SingleSubjectDesign(conditions = Dict(:drift_rate => [3, 5], :condition => [1]))
 design_seq = SequenceDesign(design_single, "SCR_")
 design_rep = RepeatDesign(design_seq, 500)
 
@@ -48,21 +45,19 @@ design_rep = RepeatDesign(design_seq, 500)
 # As parameter for the models drift_rate we use the drift_rate specified in the design by directly naming the condition as string.
 v = "drift_rate" # Drift Rate from the design
 # Now we retrieve the models default parameters, but we change the drift_rate and
-kelly_model = KellyModel(drift_rate=v, motor_onset=0.4, event_onset=0.2)
+kelly_model = KellyModel(drift_rate = v, motor_onset = 0.4, event_onset = 0.2)
 kelly_model_parameter = UnfoldSim.create_kelly_parameters_dict(kelly_model)
-drift = Drift_Component(
-    simulate_component,
-    time_vec,
-    Δt,
-    KellyModel,
-    kelly_model_parameter)
+drift = Drift_Component(simulate_component, time_vec, Δt, KellyModel, kelly_model_parameter)
 components = Dict('S' => [p3], 'C' => [drift], 'R' => [resp])
 # ## Create the Onset Component to simulate the overlap
 # To simulate an overlap between the drift_component in 'C' and the response component in 'R'. We have to specify the UniformOnset of the 'C' Component therefore with an negative offset to produce the overlap.
 seq_onset = SequenceOnset(
-    Dict('S'=>UniformOnset(width=0,offset=1.2*fs),
-         'C'=>(DriftOnset(), UniformOnset(width=0, offset=-100)),
-         'R'=>UniformOnset(width=0,offset=2*fs)))
+    Dict(
+        'S' => UniformOnset(width = 0, offset = 1.2 * fs),
+        'C' => (DriftOnset(), UniformOnset(width = 0, offset = -100)),
+        'R' => UniformOnset(width = 0, offset = 2 * fs),
+    ),
+)
 
 # ## Simulate data with the created design, components and onsets
 data, evts = UnfoldSim.simulate(
@@ -70,7 +65,7 @@ data, evts = UnfoldSim.simulate(
     design_rep,
     components,
     seq_onset,
-    NoNoise() # PinkNoise(noiselevel=1)
+    NoNoise(), # PinkNoise(noiselevel=1)
 )
 # ## Plot EEG to see the overlap as it appears
 lines(data)
@@ -82,16 +77,27 @@ current_figure()
 
 # ## Plot ERP with the overlap in the 'R' Component
 evts.event = string.(evts.event)
-data_epochs, times_epoch = Unfold.epoch(data = data, tbl = evts, τ = (-0.1, 0.9), sfreq = fs);
[…]
