+++ docs/literate/HowTo/driftComponent.jl
@@ -16,7 +16,7 @@ using CairoMakie, UnfoldMakie
 using SequentialSamplingModels
 
 fs = 500
-Δt = 1/fs; # time step
+Δt = 1 / fs; # time step
 tEnd = 1.0 # trial Duration
 time_vec = 0:Δt:tEnd; # time base - let's make it a typical stimulus duration
 # ```@raw html
@@ -25,19 +25,11 @@ time_vec = 0:Δt:tEnd; # time base - let's make it a typical stimulus duration
 
 # ## Design
 # Let's generate a single design
-design_single = SingleSubjectDesign(
-    conditions = Dict(:condition => [1])
-    )
+design_single = SingleSubjectDesign(conditions = Dict(:condition => [1]))
 # Now we convert the SingleSubjectDesign to an SequenceDesign with a Sequence of S: stimulus, C: component, R: response, _: gap
-design_seq = SequenceDesign(
-    design_single,
-    "SCR_"
-    )
+design_seq = SequenceDesign(design_single, "SCR_")
 # On top we can now repeat the Sequence to have multiple trials. For this experiment we use 100 replicas.
-design_rep = RepeatDesign(
-    design_seq,
-    100
-    )
+design_rep = RepeatDesign(design_seq, 100)
 
 # ## Create the 3 components we want to use
 # First the stimulus component as simple LinearModelComponent
@@ -57,13 +49,8 @@ v = [0.8] # Drift Rate
 A = 0.1 # The maximum start point, an indicator of an an initial bias towards a decision.
 k = 0.4 # A + k = b, where b is the decision threshold.
 t = 0.2 # The duration for a non-decisional processes (encoding and response execution).
-lba_parameter = Dict(:ν=>v, :A=>A, :k=>k, :τ=>t)
-drift = Drift_Component(
-    simulate_component,
-    time_vec,
-    Δt,
-    LBA,
-    lba_parameter)
+lba_parameter = Dict(:ν => v, :A => A, :k => k, :τ => t)
+drift = Drift_Component(simulate_component, time_vec, Δt, LBA, lba_parameter)
 # As last step we have to specify the components as a Dict connection the components with the events of the design.
 components = Dict('S' => [p3], 'C' => [drift], 'R' => [resp])
 
@@ -71,9 +58,12 @@ components = Dict('S' => [p3], 'C' => [drift], 'R' => [resp])
 # For the stimulus and response we use a simple Uniform onset with a distance of 250 and 300. The Drift_Component uses the special DriftOnset in combination with an UniformOnset. The DriftOnset returns the response time after the model reaches the decision threshold.
 # Important is to know that, the onset for each component defines the onset for the next. So in this case: S->C, C->R, R->S.
 seq_onset = SequenceOnset(
-    Dict('S'=>UniformOnset(width=0,offset=250),
-         'C'=>(DriftOnset(), UniformOnset(width=0, offset=150)),
-         'R'=>UniformOnset(width=0,offset=300)))
+    Dict(
+        'S' => UniformOnset(width = 0, offset = 250),
+        'C' => (DriftOnset(), UniformOnset(width = 0, offset = 150)),
+        'R' => UniformOnset(width = 0, offset = 300),
+    ),
+)
 
 # ## Simulate data with the created design, components and onsets
 data, evts = UnfoldSim.simulate(
@@ -81,7 +71,7 @@ data, evts = UnfoldSim.simulate(
     design_rep,
     components,
     seq_onset,
-    NoNoise() # PinkNoise(noiselevel=1)
[…]
