### 2025-02-25 6d0e09e
Apply suggestions from code review
Formatting

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
