### 2025-03-13 #6641 @danielskatz
I've suggest a few small changes in https://github.com/unfoldtoolbox/UnfoldSim.jl/pull/148 - please merge this, or let me know what you disagree with, then we can continue to publication
