### 2025-03-14 #6641 @jschepers
Thanks a lot to everyone for your patience and all the useful suggestions and comments! Your input greatly improved the package, and I learned a lot throughout the process. 🎉
